Data.Char.Block

Description

Unicode has 2-by-2 blocks, this module aims to make it more convenient to render such blocks.

Datastructures to store the state of the frame.

data Row a

A data type that determines the state of the row in a block. it determines the left and the right part of the row of the block.

Constructors

Row

Fields

left :: a

The left part of a row of the block.

right :: a

The right part of the row of the block.

rowValue

Arguments

:: Row Bool
The given Row of Bools to convert.

-> Int
The corresponding numerical value.

Convert the given Row of Booleans to an Int where the left Bool has value 1, and the right one has value two. The four different Rows thus are mapped to integers from zero to three (both inclusive).

toRow

Arguments

:: Int
The given number to convert.

-> Maybe (Row Bool)
The corresponding Row of Bools.

Convert the given number to a Row of Bools wrapped in a Just. if the value is out of bounds, Nothing is returned.

toRow'

Arguments

:: Int
The given number to convert.

-> Row Bool
The corresponding Row of Bools.

Convert the given number to a Row of Bools. If the value is out of bounds, it is unspecified what will happen.

pattern EmptyRow :: Row Bool

A pattern synonym for a Row where both the left and right subpart are False.

pattern FullRow :: Row Bool

A pattern synonym for a Row where both the left and right subpart are True.

pattern LeftRow :: Row Bool

A pattern synonym for a Row where the left part is set to True, and the right part is set to False.

pattern RightRow :: Row Bool

A pattern synonym for a Row where the left part is set to False, and the right part is set to True.

data Block a

A data type that determines the state of the four subparts of the block.

Constructors

Block

Fields

upper :: Row a

The upper part of the block.

lower :: Row a

The lower part of the block.

A unicode character that is (partially) filled block.

filled

Arguments

:: Block Bool
The given Block of Bools to convert to a Character.

-> Char
The equivalent Unicode Character for the given Block of Bools.

Convert the given Block value to a block character in unicode. True means that part is filled, and False means the part is not filled.

Convert a Character to a (partially) filled block.

fromBlock

Arguments

:: Char
The given Character to convert to a Block of Bools.

-> Maybe (Block Bool)

Convert the given Character to a Block of Bools wrapped in a Just if it exists; Nothing otherwise.

fromBlock'

Arguments

:: Char
The given Character to convert to a Block of Bools.

-> Block Bool
The equivalent Block of Bools.

Convert the given Character to a Block of Bools if it exists; unspecified result otherwise.

Pattern synonyms for blocks

pattern EmptyBlock :: Char

A pattern synonym for a block Character that will render an empty block, this is equivalent to a space.

pattern FullBlock :: Char

A pattern synonym for the block Character that will render a full block.

pattern LeftHalfBlock :: Char

A pattern synonym for a block Character that will render a block where the left half of the block is filled.

pattern RightHalfBlock :: Char

A pattern synonym for a block Character that will render a block where the right half of the block is filled.