Data.Char.Block.Sextant

Description

Unicode has 3-by-2 blocks, this module aims to make it more convenient to render such blocks.

Datastructures to store the state of the sextant.

data Sextant a

A data type that determines the state of the four subparts of the block.

Constructors

Sextant

Fields

upper :: Row a

The upper part of the sextant.

middle :: Row a

The middle part of the sextant.

lower :: Row a

The lower part of the sextant.

isSextant

Arguments

:: Char
The given Character to test.

-> Bool
True if the given Character is a sextant Character; otherwise False.

Check if the given Character is a Character that maps on a Sextant value.

A unicode character that is (partially) filled sextant.

filled

Arguments

:: Sextant Bool
The given Sextant of Bools to convert to a Char.

-> Char
The corresponding Character that presents the sextant.

Convert the given Sextant of Bools to a Char where raster items of the Sextant are written in black, and the rest in white.

Convert a Character to a (partially) filled sextant.

fromSextant

Arguments

:: Char
The Char we wish to convert to a Sextant object.

-> Maybe (Sextant Bool)
The corresponding Sextant object wrapped in a Just; Nothing if the given Char is not a sextant character.

Convert the given Char to the corresponding Sextant object wrapped in a Just data constructor. If the given Char is not a sextant character, Nothing is returned.

fromSextant'

Arguments

:: Char
The Char we wish to convert to a Sextant object.

-> Sextant Bool
The corresponding Sextant; unspecified behavior if the given Char is not a sextant character.

Convert the given Char to the corresponding Sextant object wrapped If the given Char is not a sextant character, it is unspecified what will happen.