Data.Char.Brackets

Description

Unicode considers 60 characters to be brackets: brackets are organized in pairs: each opening bracket has a corresponding closing bracket and vice versa.

The following characters are considered brackets where the first character is closed by the last character, the second by the last but one, etc.:

([{༺༼᚛⁅⁽₍⌈⌊〈❨❪❬❮❰❲❴⟅⟦⟨⟪⟬⟮⦃⦅⦇⦉⦋⦍⦏⦑⦓⦕⦗⧘⧚⧼⸢⸤⸦⸨〈《「『【〔〖〘〚﹙﹛﹝（［｛｟｢｣｠｝］）﹞﹜﹚〛〙〗〕】』」》〉⸩⸧⸥⸣⧽⧛⧙⦘⦖⦔⦒⦎⦐⦌⦊⦈⦆⦄⟯⟭⟫⟩⟧⟆❵❳❱❯❭❫❩〉⌋⌉₎⁾⁆᚜༽༻}])

These characters span over several code blocks.

Listing and converting brackets

bracketMaps :: [(Char, Char)]

A list of 2-tuples where the first item of each tuple is the opening bracket, and the second item its closing counterpart.

brackets

Arguments

:: [Char]
The list of all Chars that are brackets.

The list of all brackets characters.

openBrackets

Arguments

:: [Char]
The list of all Chars that are opening brackets.

A list of Chars that contains all opening brackets.

closeBrackets

Arguments

:: [Char]
The list of all Chars that are closing brackets.

A list of Chars that contains all closing brackets.

toOpen :: Map Char Char

A Map that maps the given close bracket characters to the corresponding open bracket.

toClose :: Map Char Char

A Map that maps the given open bracket characters to the corresponding close bracket.

Check the given bracket type

data BracketType

A data type that is used to specify the type of bracket.

Constructors

Open
The bracket is used to "open" a context.

Close
The bracket is used to "close" a context.

isBracket

Arguments

:: Char
The given Char to test.

-> Bool
True if the given Char is an open bracket; False otherwise.

Check if the given Char is a bracket character.

bracketType :: Char -> Maybe BracketType

Check the BracketType of the Char wrapped in a Just data construct; Nothing if the given Char is not a bracket character.

bracketType' :: Char -> BracketType

Check the BracketType of the Char. For a Char that is not a bracket the behavior is unspecified.

isOpenBracket

Arguments

:: Char
The given Char to test.

-> Bool
True if the Char is an open bracket; False otherwise.

Check if the given Char is an open bracket.

isCloseBracket

Arguments

:: Char
The given Char to test.

-> Bool
True if the Char is an close bracket; False otherwise.

Check if the given Char is a close bracket.

Determine the opposite bracket

getOppositeChar

Arguments

:: Char
The given Char for which we want to determine the opposite bracket.

-> Maybe Char
The opposite bracket wrapped in a Just if the given Char is a bracket character; Nothing otherwise.

Get the bracket character that is the counterpart of the given bracket character wrapped in a Just data constructor. If the given Char is not a bracket, Nothing is returned.

getOppositeChar'

Arguments

:: Char
The given Char for which we want to determine the opposite bracket.

-> Char
The opposite bracket if the given Char is a bracket; otherwise the given Char.

Get the bracket character that is the counterpart of the given bracket character. If the given Char is not a bracket, the given Char is returned.