Data.Char.Card

Description

The 1F0A0 code block can be used to work with cards. This module makes working with the three sets of cards more convenient. The first set are cards with a suit (four suits) and a rank (14 ranks), which allows us to generate 56 cards. Next there are three jokes with the colors red, black and white. Finally there are 21 trump cards and the fool card (a tarot card which normally has no number).

The module contains pattern synonyms that define common aliasses for these cards.

Data structures to define a card

data CardSuit

A data type for the card suits

Constructors

Spades
The spades card suit.

Hearts
The hearts card suit.

Diamonds
The diamonds card suit.

Clubs
The clubs card suit.

data CardRank

A data type for the rank of the card.

Constructors

Ace
The ace card rank.

R2
Card rank 2.

R3
Card rank 3.

R4
Card rank 4.

R5
Card rank 5.

R6
Card rank 6.

R7
Card rank 7.

R8
Card rank 8.

R9
Card rank 9.

R10
Card rank 10.

Jack
The jack card rank.

Knight
The knight card rank.

Queen
The queen card rank.

King
The king card rank.

data JokerColor

A data type to represent the three colors for which there are jokers: red, black and white.

Constructors

Red
The red joker.

Black
The black joker.

White
The white joker.

data Trump

A data type for the trump cards, often used for tarot.

Constructors

Fool
The fool trump card, this tarot card is normally not numbered.

Trump1
Tarot card I.

Trump2
Tarot card II.

Trump3
Tarot card III.

Trump4
Tarot card IV.

Trump5
Tarot card V.

Trump6
Tarot card VI.

Trump7
Tarot card VII.

Trump8
Tarot card VIII.

Trump9
Tarot card IX.

Trump10
Tarot card X.

Trump11
Tarot card XI.

Trump12
Tarot card XII.

Trump13
Tarot card XIII.

Trump14
Tarot card XIV.

Trump15
Tarot card XV.

Trump16
Tarot card XVI.

Trump17
Tarot card XVII.

Trump18
Tarot card XVIII.

Trump19
Tarot card XIX.

Trump20
Tarot card XX.

Trump21
Tarot card XXI.

data Card

A data type that represents the possible types of cards for which there is a Unicode characters. This is the back of a card, a card with a suit and rank, three jokers, and the 21 trump cards and the fool.

Constructors

Back
The back of the card.

Card CardSuit CardRank
A card that is a combination of a CardSuit and a CardRank. There are 56 possibilities.

Joker JokerColor
Three possible JokerColor cards.

Trump Trump

Converting cards to the corresponding Unicode character

back :: Char

The unicode character that represents the back of the card.

card :: Card -> Char

Convert the given Card object to its unicode equivalent.

card'

Arguments

:: CardSuit
The given CardSuit for the card.

-> CardRank
The given CardRank for the card.

-> Char
The corresponding unicode character with the given suit and rank.

Convert the given CardSuit and CardRank to the equivalent unicode character for this card.

joker

Arguments

:: JokerColor
The given JokerColor to convert to a unicode character.

-> Char
The unicode character that represents the joker with the given color.

Convert the given JokerColor to the unicode character which represents this joker color.

trump

Arguments

:: Trump
The given 't:Trump' value to convert to a unicode character.

-> Char
The unicode character that represents the given 't:Trump' playing card.

Convert the given 't:Trump' value to the unicode equivalent.

Pattern synonyms for cards

Aliasses for the card suits

pattern Swords :: CardSuit

Swords is an alias for the spades card suit.

pattern Cups :: CardSuit

Cups is an alias for the hearts card suit.

pattern Pentacles :: CardSuit

Pentacles is an alias for the diamonds card suit.

pattern Wands :: CardSuit

Wands is an alias for the clubs card suit.

Aliasses for the card ranks

Aliasses for the jack

pattern Valet :: CardRank

In France, the jack is sometimes called the valet.

pattern Bube :: CardRank

In Germany, Austria and Switzerland, the jack is sometimes called the bube.

pattern Unter :: CardRank

In Germany and Switzerland, the jack is sometimes called the unter.

pattern Page :: CardRank

An alternative name for the jack is page.

pattern Fante :: CardRank

In Italy, the jack is sometimes called the fante.

Aliasses for the knight

pattern Chevalier :: CardRank

In France, the knight is sometimes called the chevalier.

pattern Ober :: CardRank

In Germany, the knight is sometimes called the ober.

pattern Ritter :: CardRank

In Germany, the knight is sometimes called the ritter.

pattern Cavall :: CardRank

An alternative name for the jack is cavall.

pattern Cavaliere :: CardRank

An alternative name for the jack is cavaliere.

Aliasses for the queen

pattern Dame :: CardRank

An alternative name for the queen is dame.

pattern Königin :: CardRank

In Germany, the queen is sometimes called the königin.

pattern Regina :: CardRank

In Italy, the queen is sometimes called the regina.

Aliasses for the king

pattern Roi :: CardRank

In France, the king is sometimes called the roi.

pattern König :: CardRank

In Germany, the king is sometimes called the könig.

pattern Re :: CardRank

In Italy, the queen is sometimes called the re.

Trump patterns

The four ages

pattern Childhood :: Trump

The trump card with number II is named childhood.

pattern Youth :: Trump

The trump card with number III is named youth.

pattern Maturity :: Trump

The trump card with number IV is named maturity.

pattern OldAge :: Trump

The trump card with number V is named old age.

The four times of the day

pattern Morning :: Trump

The trump card with number VI is named morning.

pattern Afternoon :: Trump

The trump card with number VII is named afternoon.

pattern Evening :: Trump

The trump card with number VIII is named evening.

pattern Night :: Trump

The trump card with number IX is named night.

The four elements

pattern Earth :: Trump

The trump card with number X is named earth.

pattern Air :: Trump

The trump card with number X is named air.

pattern Water :: Trump

The trump card with number XI is named water.

pattern Fire :: Trump

The trump card with number XI is named fire.

The four leisures

pattern Dance :: Trump

The trump card with number XII is named dance.

pattern Shopping :: Trump

The trump card with number XIII is named shopping.

pattern OpenAir :: Trump

The trump card with number XIV is named open air.

pattern VisualArts :: Trump

The trump card with number XV is named visual arts.

The four seasons

pattern Spring :: Trump

The trump card with number XVI is named spring.

pattern Summer :: Trump

The trump card with number XVII is named summer.

pattern Autumn :: Trump

The trump card with number XVIII is named autumn.

pattern Winter :: Trump

The trump card with number XIX is named winter.

The game

pattern Game :: Trump

The trump card with number XX is named the game.

Folly

pattern Collective :: Trump

The trump card with number XXI is named collective.

pattern Individual :: Trump

The trump card with number I is named individual.