Data.Char.Core

Description

This module defines data structures that are used in other modules, for example to rotate the characters.

Possible rotations

data Orientation

The possible orientations of a unicode character, these can be horizontal, or vertical.

Constructors

Horizontal
Horizontal orientation.

Vertical
Vertical orientation.

data Rotate90

Possible rotations of a unicode character if that character can be rotated over 0, 90, 180, and 270 degrees.

Constructors

R0
No rotation.

R90
Rotation over 90 degrees.

R180
Rotation over 180 degrees.

R270
Rotation over 270 degrees.

Rotated objects

data Oriented a

A data type that specifies that an item has been given an orientation.

Constructors

Oriented

Fields

oobject :: a

The object that is oriented.

orientation :: Orientation

The oriented of the oriented object.

data Rotated a

A data type that specifies that an item has been given a rotation.

Constructors

Rotated

Fields

robject :: a

The object that is rotated.

rotation :: Rotate90

The rotation of the rotated object.

Letter case

data LetterCase

Specify whether we write a value in UpperCase or LowerCase. The Default is UpperCase, since for example often Roman numerals are written in upper case.

Constructors

UpperCase
The upper case formatting.

LowerCase
The lower case formatting.

splitLetterCase

Arguments

:: a
The value to return in case of UpperCase.

-> a
The value to return in case of LowerCase.

-> LetterCase
The given letter case.

-> a
One of the two given values, depending on the LetterCase value.

Pick one of the two values based on the LetterCase value.

Ligating

data Ligate

Specify if one should ligate, or not. When litigation is done characters that are normally written in two (or more) characters are combined in one character. For example Ⅲ instead of ⅠⅠⅠ.

Constructors

Ligate
A ligate operation is performed on the characters, the def for 't:Ligate'.

NoLigate
No ligate operation is performed on the charaters.

splitLigate

Arguments

:: a
The value to return in case of 'v:Ligate'.

-> a
The value to return in case of NoLigate.

-> Ligate
The ligation style.

-> a
One of the two given values, based on the 't:Ligate' value.

Pick one of the two values based on the value for 't:Ligate'.

ligate :: (a -> a) -> Ligate -> a -> a

Specify if the given ligate function should be performed on the input, if 'v:Ligate' is passed, and the identity function otherwise.

ligateF :: Functor f => (a -> a) -> Ligate -> f a -> f a

Specify if the given ligate function is performed over the functor object if 'v:Ligate' is passed, and the identity function otherwise.

Types of fonts

data Emphasis

A data type that lists the possible emphasis of a font. This can be Bold or NoBold the Default is NoBold.

Constructors

NoBold
The characters are not stressed with boldface.

Bold
The characters are stressed in boldface.

splitEmphasis

Arguments

:: a
The value to return in case of NoBold.

-> a
The value to return in case of Bold.

-> Emphasis
The emphasis type.

-> a
One of the two given values, based on the 't:Emphasis' value.

Pick one of the two values based on the 't:Emphasis' value.

data ItalicType

A data type that can be used to specify if an italic character is used. The Default is NoItalic.

Constructors

NoItalic
No italic characters are used.

Italic
Italic characters are used.

splitItalicType

Arguments

:: a
The value to return in case of NoItalic.

-> a
The value to return in case of Italic.

-> ItalicType
The italic type.

-> a
One of the two given values, based on the 't:ItalicType' value.

Pick one of the two values based on the 't:ItalicType' value.

data FontStyle

A data type that specifies if the font is with serifs or not. The 'Defaul;t' is Serif.

Constructors

SansSerif
The character is a character rendered without serifs.

Serif
The character is a character rendered with serifs.

splitFontStyle

Arguments

:: a
The value to return in case of SansSerif.

-> a
The value to return in case of Serif.

-> FontStyle
The font style.

-> a
One of the two given values, based on the 't:FontStyle' value.

Pick one of the two values based on the 't:FontStyle' value.

Character range checks

isAsciiAlphaNum :: Char -> Bool

Checks if a character is an alphabetic or numerical character in ASCII. The characters 0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz satisfy this predicate.

isAsciiAlpha :: Char -> Bool

Checks if a charcter is an alphabetic character in ASCII. The characters ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz satisfy this predicate.

isGreek :: Char -> Bool

Checks if a character is a basic greek alphabetic character or a Greek-like symbol. The characters ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ satisfy this predicate.

isACharacter

Arguments

:: Char
The given Character to check.

-> Bool
True if the given Character is a character (according to the Unicode specifications); False otherwise.

Check if the given character is a character according to the Unicode specifications. Codepoints that are not a character are denoted in the Unicode documentation with <not a character>.

isNotACharacter

Arguments

:: Char
The given Character to check.

-> Bool
True if the given Character is not a character (according to the Unicode specifications); False otherwise.

Check if the given character is not a character according to the Unicode specifications. The Unicode documentation denotes these with <not a character>.

isReserved

Arguments

:: Char
The given Character to check.

-> Bool
True if the given Character is reserved; False otherwise.

Check if the given character is a reserved character. This is denoted in the Unicode documentation with <reserved>.

isNotReserved

Arguments

:: Char
The given Character to check.

-> Bool
True if the given Character is not reserved; False otherwise.

Check if the given character is not a reserved character. This is denoted in the Unicode documentation with <reserved>.

Map characters from and to Enums

mapFromEnum

Arguments

:: Enum a

=> Int
The given offset value.

-> a
The given Enum value to convert to a Character.

-> Char
The character that corresponds to the given Enum object.

Map the given object with a type that is an instance of Enum to a Character with a given offset for the Character value.

mapToEnum

Arguments

:: Enum a

=> Int
The given offset value.

-> Char
The Character to map to an Enum object.

-> a
The given Enum object for the given Char.

Map the given Char object to an object with a type that is an instance of Enum with a given offset for the Character range.

mapToEnumSafe

Arguments

:: forall a. (Bounded a, Enum a)

=> Int
The given offset value.

-> Char
The given Character to map to an Enum object.

-> Maybe a
The given Enum object for the given Character wrapped in a Just if that exists; Nothing otherwise.

Map the given Char object to an object with a type that is an instance of Enum. It first checks if the mapping results in a value between the fromEnum values for minBound and maxBound.

liftNumberFrom

Arguments

:: Int
The given offset value.

-> Int
The maximum value that can be mapped.

-> Int
The given Unicode value used for the offset.

-> Int
The given number to convert, must be between the offset and the maximum.

-> Maybe Char
The corresponding Character wrapped in a Just if the number is between the offset and the maximum; Nothing otherwise.

Construct a function that maps digits to the character with the given value for the offset.

liftNumberFrom'

Arguments

:: Int
The given offset value.

-> Int
The given Unicode value used for the offset.

-> Int
The given number to convert to a corresponding Character.

-> Char
The corresponding Character for the given mapping function.

Construct a function that maps digits to the character with the given value for the offset.

liftNumber

Arguments

:: Int
The maximum value that can be mapped.

-> Int
The given Unicode value used for 0.

-> Int
The given digit to convert to a number between 0 and the maximum.

-> Maybe Char
The corresponding Character wrapped in a Just if the number is between 0 and 9; Nothing otherwise.

Construct a function that maps digits to the character with the given value for 0.

liftNumber'

Arguments

:: Int
The given Unicode value used for 0.

-> Int
The given digit to convert.

-> Char
The corresponding Character, for numbers outside the 0-9 range, the result is unspecified.

Construct a function that maps digits to characters with the given value for 0.

liftDigit

Arguments

:: Int
The given Unicode value used for 0.

-> Int
The given digit to convert to a number between 0 and 9.

-> Maybe Char
The corresponding Character wrapped in a Just if the number is between 0 and 9; Nothing otherwise.

Construct a function that maps digits to the character with the given value for 0.

liftDigit'

Arguments

:: Int
The given Unicode value used for 0.

-> Int
The given digit to convert, must be between 0 and 9.

-> Char
The corresponding Character, for numbers outside the 0-9 range, the result is unspecified.

Construct a function that maps digits to characters with the given value for 0.

liftUppercase

Arguments

:: Int
The given Unicode value for A.

-> Char
The given character to convert.

-> Maybe Char
The corresponding character wrapped in a Just if the given character is in the A-Z range; Nothing otherwise.

Construct a function that maps upper case alphabetic characters with the given value for A.

liftUppercase'

Arguments

:: Int
The given Unicode value for A.

-> Char
The given upper case alphabetic value to convert.

-> Char
The corresponding character, if the given value is outside the A-Z range, the result is unspecified.

Construct a function that maps upper case alphabetic characters with the given value for A.

liftLowercase

Arguments

:: Int
The given Unicode value for a.

-> Char
The given character to convert.

-> Maybe Char
The corresponding character wrapped in a Just if the given character is in the a-z range; Nothing otherwise.

Construct a function that maps lower case alphabetic characters with the given value for a.

liftLowercase'

Arguments

:: Int
The given Unicode value for a.

-> Char
The given upper case alphabetic value to convert.

-> Char
The corresponding character, if the given value is outside the a-z range, the result is unspecified.

Construct a function that maps lower case alphabetic characters with the given value for a.

liftUpperLowercase

Arguments

:: Int
The given Unicode value for A.

-> Int
The given Unicode value for a.

-> Char
The given character to convert.

-> Maybe Char
The corresponding character wrapped in a Just if the given character is in the A-Z,a-z range; Nothing otherwise.

Construct a function that maps lower case alphabetic characters with the given values for A and a.

liftUpperLowercase'

Arguments

:: Int
The given Unicode value for A.

-> Int
The given Unicode value for a.

-> Char
The given character to convert.

-> Char
The corresponding character if the given character is in the A-Z,a-z range; unspecified otherwise.

Construct a function that maps lower case alphabetic characters with the given values for A and a.

Convert objects from and to Unicode Characters

class UnicodeCharacter a where

A class from which objects can be derived that map to and from a single unicode character.

Minimal complete definition

toUnicodeChar, fromUnicodeChar

Methods

toUnicodeChar

Arguments

:: a
The given object to convert to a Character.

-> Char
The equivalent Unicode Character.

Convert the given object to a Unicode Character.

fromUnicodeChar

Arguments

:: Char
The given Character to convert to an element.

-> Maybe a
An element if the given Character maps to an element wrapped in a Just; Nothing otherwise.

Convert the given Character to an object wrapped in a Just data constructor if that exists; Nothing otherwise.

fromUnicodeChar'

Arguments

:: Char
The given Character to convert to an element.

-> a
The given element that is equivalent to the given Character.

Convert the given Character to an object. If the Character does not map on an element, the behavior is unspecified, it can for example result in an error.

isInCharRange

Arguments

:: Char
The given Character to test.

-> Bool
True if the given Character has a corresponding value for a; False otherwise.

Check if the given Character maps on an item of a.

type UnicodeChar = UnicodeCharacter

An alias of the UnicodeCharacter type class.

class UnicodeText a where

A class from which boejcts can be derived that map to and from a sequence of unicode characters.

Minimal complete definition

Nothing

Methods

toUnicodeText

Arguments

:: a
The given object to convert to a Text object.

-> Text
A Text object that is the Unicode representation of the element.

Convert the given object to a Text object.

default toUnicodeText :: UnicodeCharacter a => a -> Text

fromUnicodeText

Arguments

:: Text
The given Text to convert to an object.

-> Maybe a
The equivalent object wrapped in a Just data constructor if it exists; Nothing otherwise.

Convert the given Text to an object wrapped in a Just data constructor if that exists; Nothing otherwise.

default fromUnicodeText :: UnicodeCharacter a => Text -> Maybe a

fromUnicodeText'

Arguments

:: Text
The given Text to convert to an object.

-> a
The given equivalent object. If there is no equivalent object, the behavior is unspecified.

Convert the given Text to an object. If the Text does not map on an element, the behavior is unspecified, it can for example result in an error.

isInTextRange

Arguments

:: Text
The given Text object to test.

-> Bool
True if there is a counterpart of type a; False otherwise.

Determine if the given Text value maps on a value of type a.

generateIsInTextRange

Arguments

:: (Char -> Bool)
The given isInCharRange check.

-> Text
The Text object to check.

-> Bool
True if the given Text object has a single character for which the isInCharRange check succeeds, False otherwise.

Convert a given isInCharRange check into a isInTextRange check.

generateIsInTextRange'

Arguments

:: forall a. UnicodeCharacter a

=> Text
The given Text object to check.

-> Bool
True if the given Text object has a single character for which the isInCharRange check succeeds, False otherwise.

Generate an isInTextRange check with the isInCharRange check for the instance of UnicodeCharacter of that type.

Mirroring items horizontally and/or vertically

class MirrorHorizontal a where

A type class that specifies that the items can be mirrored in the horizontal direction (such that up is now down). The mirror is not per se pixel perfect. For example the mirror of 🂁 is 🁵, so the dots of the bottom pat of the domino are not mirrored correctly.

Methods

mirrorHorizontal

Arguments

:: a
The given item to mirror horizontally.

-> a
The corresponding mirrored item.

Obtain the horizontally mirrored variant of the given item. Applying the same function twice should return the original object.

class MirrorVertical a where

A type class that specifies that the items can be mirrored in the vertical direction (such that left is now right). The mirror is not per se pixel perfect. For example the vertical mirror of 🁏 is 🁃, so the dots of the right part of the domino are not mirrored correctly.

Methods

mirrorVertical

Arguments

:: a
The given item to mirror vertically.

-> a
The corresponding mirrored item.

Obtain the vertically mirrored variant of the given item. Applying the same function twice should return the original object.

Ways to display numbers

data PlusStyle

Specify whether we write a positive number with or without a plus sign. the Default is WithoutPlus.

Constructors

WithoutPlus
Write positive numbers without using a plus sign.

WithPlus
Write positive numbers with a plus sign.

splitPlusStyle

Arguments

:: a
The value to return in case of WithoutPlus.

-> a
The value to return in case of WithPlus.

-> PlusStyle
The plus style.

-> a
One of the two given values, based on the 't:PlusStyle' value.

Pick one of the two values based on the 't:PlusStyle' value.

Functions to implement a number system

withSign

Arguments

:: Integral i

=> (i -> Text)
The function that maps the absolute value of the number to a Text object that is appended to the sign.

-> Char
The plus sign to use.

-> Char
The minus sign to use.

-> PlusStyle
The given PlusStyle to use.

-> i
The given Integral number to render.

-> Text
A Text object that represents the given number, with the given sign numbers in the given PlusStyle.

Calculate for a given plus and minus sign a Text object for the given number in the given PlusStyle.

signValueSystem

Arguments

:: Integral i

=> i
The given radix to use.

-> (Int -> Int -> Text)
A function that maps the value and the weight to a Text object.

-> Text
The given Text used to represent zero.

-> Char
The given Char used to denote plus.

-> Char
The given Char used to denote minus.

-> PlusStyle
The given PlusStyle to use.

-> i
The given number to convert.

-> Text
A Text object that denotes the given number with the given sign-value system.

A function to make it more convenient to implement a sign-value system. This is done for a given radix a function that maps the given value and the given weight to a Text object, a Text object for zero (since in some systems that is different), and characters for plus and minus. The function then will for a given PlusStyle convert the number to a sequence of characters with respect to how the sign-value system is implemented.

positionalNumberSystem

Arguments

:: Integral i

=> i
The given radix to use.

-> (Int -> Char)
A function that maps the value of a digit to the corresponding Char.

-> Char
The given character used to denote plus.

-> Char
The given character used to denote minus.

-> PlusStyle
The given PlusStyle to use.

-> i
The given number to convert.

-> Text
A Text object that denotes the given number with the given positional number system.

A function to make it more convenient to implement a /positional number system. This is done for a given radix/ a given conversion funtion that maps a value to a Char, and a Char for plus and minus. The function then construct a Text object for a given PlusStyle and a given number.

positionalNumberSystem10

Arguments

:: Integral i

=> (Int -> Char)
A function that maps the value of a digit to the corresponding Char.

-> Char
The given character used to denote plus.

-> Char
The given character used to denote minus.

-> PlusStyle
The given PlusStyle to use.

-> i
The given number to convert.

-> Text
A Text object that denotes the given number with the given positional number system.

A function to make it more convenient to implement a /positional number system with radix/ 10.

Re-export of some functions of the Char module

chr :: Int -> Char

The toEnum method restricted to the type Char.

isAlpha :: Char -> Bool

Selects alphabetic Unicode characters (lower-case, upper-case and title-case letters, plus letters of caseless scripts and modifiers letters). This function is equivalent to isLetter.

isAlphaNum :: Char -> Bool

Selects alphabetic or numeric Unicode characters.

Note that numeric digits outside the ASCII range, as well as numeric characters which aren't digits, are selected by this function but not by isDigit. Such characters may be part of identifiers but are not used by the printer and reader to represent numbers.

isAscii :: Char -> Bool

Selects the first 128 characters of the Unicode character set, corresponding to the ASCII character set.

ord :: Char -> Int

The fromEnum method restricted to the type Char.