Data.Char.Currency

Description

Unicode has multiple code blocks where it defines currencies. This module aims to expose a data structure that makes it more convenient to work with currency characters.

Defining currencies

data Currency

A datatype to present the currencies that have a unicode character.

Constructors

Dollar
A currency that is rendered as $.

Cent
A currency that is rendered as ¢.

Pound
A currency that is rendered as £.

Currency
A currency that is rendered as ¤.

Yen
A currency that is rendered as ¥.

ArmenianDram
A currency that is rendered as ֏.

Afghani
A currency that is rendered as ؋.

NkoDorome
A currency that is rendered as ߾.

NkoTaman
A currency that is rendered as ߿.

BengaliRupeeMark
A currency that is rendered as ৲.

BengaliRupee
A currency that is rendered as ৳.

BengaliGandaMark
A currency that is rendered as ৻.

GujaratiRupee
A currency that is rendered as ૱.

TamilRupee
A currency that is rendered as ௹.

ThaiSymbolBaht
A currency that is rendered as ฿.

KhmerSymbolRiel
A currency that is rendered as ៛.

EuroCurrency
A currency that is rendered as ₠.

Colon
A currency that is rendered as ₡.

Cruzeiro
A currency that is rendered as ₢.

FrenchFranc
A currency that is rendered as ₣.

Lira
A currency that is rendered as ₤.

Mill
A currency that is rendered as ₥.

Naira
A currency that is rendered as ₦.

Peseta
A currency that is rendered as ₧.

Rupee
A currency that is rendered as ₨.

Won
A currency that is rendered as ₩.

NewSheqel
A currency that is rendered as ₪.

Dong
A currency that is rendered as ₫.

Euro
A currency that is rendered as €.

Kip
A currency that is rendered as ₭.

Tugrik
A currency that is rendered as ₮.

Drachma
A currency that is rendered as ₯.

GermanPenny
A currency that is rendered as ₰.

Peso
A currency that is rendered as ₱.

Guarani
A currency that is rendered as ₲.

Austral
A currency that is rendered as ₳.

Hryvnia
A currency that is rendered as ₴.

Cedi
A currency that is rendered as ₵.

LivreTournois
A currency that is rendered as ₶.

Spesmilo
A currency that is rendered as ₷.

Tenge
A currency that is rendered as ₸.

IndianRupee
A currency that is rendered as ₹.

TurkishLira
A currency that is rendered as ₺.

NordicMark
A currency that is rendered as ₻.

Manat
A currency that is rendered as ₼.

Ruble
A currency that is rendered as ₽.

Lari
A currency that is rendered as ₾.

Bitcoin
A currency that is rendered as ₿.

NorthIndicRupeeMark
A currency that is rendered as ꠸.

Rial
A currency that is rendered as ﷼.

SmallDollar
A currency that is rendered as ﹩.

FullwidthDollar
A currency that is rendered as ＄.

FullwidthCent
A currency that is rendered as ￠.

FullwidthPound
A currency that is rendered as ￡.

FullwidthYen
A currency that is rendered as ￥.

FullwidthWon
A currency that is rendered as ￦.

TamilKaacu
A currency that is rendered as 𑿝.

TamilPanam
A currency that is rendered as 𑿞.

TamilPon
A currency that is rendered as 𑿟.

TamilVaraakan
A currency that is rendered as 𑿠.

WanchoNgun
A currency that is rendered as 𞋿.

IndicSiyaqRupeeMark
A currency that is rendered as 𞲰.

Currencies as Char objects

dollar

Arguments

:: Char
A character that corresponds with the dollar sign.

The character used to render a dollar sign presented as $.

cent

Arguments

:: Char
A character that corresponds with the cent sign.

The character used to render a cent sign presented as ¢.

pound

Arguments

:: Char
A character that corresponds with the pound sign.

The character used to render a pound sign presented as £.

currency

Arguments

:: Char
A character that corresponds with the currency sign.

The character used to render a currency sign presented as ¤.

yen

Arguments

:: Char
A character that corresponds with the yen sign.

The character used to render a yen sign presented as ¥.

armenianDram

Arguments

:: Char
A character that corresponds with the armenian dram sign.

The character used to render a armenian dram sign presented as ֏.

afghani

Arguments

:: Char
A character that corresponds with the afghani sign.

The character used to render a afghani sign presented as ؋.

nkoDorome

Arguments

:: Char
A character that corresponds with the nko dorome sign.

The character used to render a nko dorome sign presented as ߾.

nkoTaman

Arguments

:: Char
A character that corresponds with the nko taman sign.

The character used to render a nko taman sign presented as ߿.

bengaliRupeeMark

Arguments

:: Char
A character that corresponds with the bengali rupee mark.

The character used to render a bengali rupee mark presented as ৲.

bengaliRupee

Arguments

:: Char
A character that corresponds with the bengali rupee sign.

The character used to render a bengali rupee sign presented as ৳.

bengaliGandaMark

Arguments

:: Char
A character that corresponds with the bengali ganda mark.

The character used to render a bengali ganda mark presented as ৻.

gujaratiRupee

Arguments

:: Char
A character that corresponds with the gujarati rupee sign.

The character used to render a gujarati rupee sign presented as ૱.

tamilRupee

Arguments

:: Char
A character that corresponds with the tamil rupee sign.

The character used to render a tamil rupee sign presented as ௹.

thaiSymbolBaht

Arguments

:: Char
A character that corresponds with the thai currency symbol baht.

The character used to render a thai currency symbol baht presented as ฿.

khmerSymbolRiel

Arguments

:: Char
A character that corresponds with the khmer currency symbol riel.

The character used to render a khmer currency symbol riel presented as ៛.

euroCurrency

Arguments

:: Char
A character that corresponds with the euro-currency sign.

The character used to render a euro-currency sign presented as ₠.

colon

Arguments

:: Char
A character that corresponds with the colon sign.

The character used to render a colon sign presented as ₡.

cruzeiro

Arguments

:: Char
A character that corresponds with the cruzeiro sign.

The character used to render a cruzeiro sign presented as ₢.

frenchFranc

Arguments

:: Char
A character that corresponds with the french franc sign.

The character used to render a french franc sign presented as ₣.

lira

Arguments

:: Char
A character that corresponds with the lira sign.

The character used to render a lira sign presented as ₤.

mill

Arguments

:: Char
A character that corresponds with the mill sign.

The character used to render a mill sign presented as ₥.

naira

Arguments

:: Char
A character that corresponds with the naira sign.

The character used to render a naira sign presented as ₦.

peseta

Arguments

:: Char
A character that corresponds with the peseta sign.

The character used to render a peseta sign presented as ₧.

rupee

Arguments

:: Char
A character that corresponds with the rupee sign.

The character used to render a rupee sign presented as ₨.

won

Arguments

:: Char
A character that corresponds with the won sign.

The character used to render a won sign presented as ₩.

newSheqel

Arguments

:: Char
A character that corresponds with the new sheqel sign.

The character used to render a new sheqel sign presented as ₪.

dong

Arguments

:: Char
A character that corresponds with the dong sign.

The character used to render a dong sign presented as ₫.

euro

Arguments

:: Char
A character that corresponds with the euro sign.

The character used to render a euro sign presented as €.

kip

Arguments

:: Char
A character that corresponds with the kip sign.

The character used to render a kip sign presented as ₭.

tugrik

Arguments

:: Char
A character that corresponds with the tugrik sign.

The character used to render a tugrik sign presented as ₮.

drachma

Arguments

:: Char
A character that corresponds with the drachma sign.

The character used to render a drachma sign presented as ₯.

germanPenny

Arguments

:: Char
A character that corresponds with the german penny sign.

The character used to render a german penny sign presented as ₰.

peso

Arguments

:: Char
A character that corresponds with the peso sign.

The character used to render a peso sign presented as ₱.

guarani

Arguments

:: Char
A character that corresponds with the guarani sign.

The character used to render a guarani sign presented as ₲.

austral

Arguments

:: Char
A character that corresponds with the austral sign.

The character used to render a austral sign presented as ₳.

hryvnia

Arguments

:: Char
A character that corresponds with the hryvnia sign.

The character used to render a hryvnia sign presented as ₴.

cedi

Arguments

:: Char
A character that corresponds with the cedi sign.

The character used to render a cedi sign presented as ₵.

livreTournois

Arguments

:: Char
A character that corresponds with the livre tournois sign.

The character used to render a livre tournois sign presented as ₶.

spesmilo

Arguments

:: Char
A character that corresponds with the spesmilo sign.

The character used to render a spesmilo sign presented as ₷.

tenge

Arguments

:: Char
A character that corresponds with the tenge sign.

The character used to render a tenge sign presented as ₸.

indianRupee

Arguments

:: Char
A character that corresponds with the indian rupee sign.

The character used to render a indian rupee sign presented as ₹.

turkishLira

Arguments

:: Char
A character that corresponds with the turkish lira sign.

The character used to render a turkish lira sign presented as ₺.

nordicMark

Arguments

:: Char
A character that corresponds with the nordic mark sign.

The character used to render a nordic mark sign presented as ₻.

manat

Arguments

:: Char
A character that corresponds with the manat sign.

The character used to render a manat sign presented as ₼.

ruble

Arguments

:: Char
A character that corresponds with the ruble sign.

The character used to render a ruble sign presented as ₽.

lari

Arguments

:: Char
A character that corresponds with the lari sign.

The character used to render a lari sign presented as ₾.

bitcoin

Arguments

:: Char
A character that corresponds with the bitcoin sign.

The character used to render a bitcoin sign presented as ₿.

northIndicRupeeMark

Arguments

:: Char
A character that corresponds with the north indic rupee mark.

The character used to render a north indic rupee mark presented as ꠸.

rial

Arguments

:: Char
A character that corresponds with the rial sign.

The character used to render a rial sign presented as ﷼.

smallDollar

Arguments

:: Char
A character that corresponds with the small dollar sign.

The character used to render a small dollar sign presented as ﹩.

fullwidthDollar

Arguments

:: Char
A character that corresponds with the fullwidth dollar sign.

The character used to render a fullwidth dollar sign presented as ＄.

fullwidthCent

Arguments

:: Char
A character that corresponds with the fullwidth cent sign.

The character used to render a fullwidth cent sign presented as ￠.

fullwidthPound

Arguments

:: Char
A character that corresponds with the fullwidth pound sign.

The character used to render a fullwidth pound sign presented as ￡.

fullwidthYen

Arguments

:: Char
A character that corresponds with the fullwidth yen sign.

The character used to render a fullwidth yen sign presented as ￥.

fullwidthWon

Arguments

:: Char
A character that corresponds with the fullwidth won sign.

The character used to render a fullwidth won sign presented as ￦.

tamilKaacu

Arguments

:: Char
A character that corresponds with the tamil sign kaacu.

The character used to render a tamil sign kaacu presented as 𑿝.

tamilPanam

Arguments

:: Char
A character that corresponds with the tamil sign panam.

The character used to render a tamil sign panam presented as 𑿞.

tamilPon

Arguments

:: Char
A character that corresponds with the tamil sign pon.

The character used to render a tamil sign pon presented as 𑿟.

tamilVaraakan

Arguments

:: Char
A character that corresponds with the tamil sign varaakan.

The character used to render a tamil sign varaakan presented as 𑿠.

wanchoNgun

Arguments

:: Char
A character that corresponds with the wancho ngun sign.

The character used to render a wancho ngun sign presented as 𞋿.

indicSiyaqRupeeMark

Arguments

:: Char
A character that corresponds with the indic siyaq rupee mark.

The character used to render a indic siyaq rupee mark presented as 𞲰.

Check if a character is a currency

isCurrency

Arguments

:: Char
The given character to test.

-> Bool
True if the given character is a currency character; False otherwise.

Check if the given Character is a currency character.