Data.Char.Dice

Description

Unicode has a block named Miscellaneous Symbols that includes unicode characters for dice, this module aims to make it more convenient to render die characters.

Represent die values

data DieValue

A data type to store the values of a die.

Constructors

I
A die with value one, represented with ⚀.

II
A die with value two, represented with ⚁.

III
A die with value three, represented with ⚂.

IV
A die with value four, represented with ⚃.

V
A die with value five, represented with ⚄.

VI
A die with value six, represented with ⚅.

Convert to a die value

toDieValue

Arguments

:: Integral i

=> i
The given integral value to convert to a DieValue.

-> Maybe DieValue
A DieValue wrapped in a Just if the given integral value is greater than zero, and less than seven, otherwise Nothing.

Convert the given integral value to a DieValue that represents the given number. If the number is less than one, or greater than six, Nothing is returned.

Render a die

die

Arguments

:: DieValue
The die value to convert.

-> Char
A unicode character that represents a die with the given DieValue.

Convert the given DieValue to a unicode character that represents a die with that value.