Data.Char.Domino

Description

A module that defines values for domino pieces, and converts these to unicode characters of the 1F030 unicode block.

Data types to represent domino values

data Domino a

A domino piece, which has two items. Depending on the orientation, the items are located at the top and bottom; or left and right.

Constructors

Domino
The front side of the domino piece.

Fields

leftTop :: a

The part that is located at the left side in case the piece is located horizontally, or at the top in case the piece is located vertically.

rightBottom :: a

The part that is located at the right side in case the piece is located horizontally, or at the bottom in case the piece is located vertically.

Back
The back side of the domino piece.

pattern (:|)

Arguments

:: a
The item that is located at the left, or the top.

-> a
The item that is located at the right, or the bottom.

-> Domino a
The domino that is constructed.

A pattern synonym that makes it more convenient to write expressions that look like domino's like for example II :| IV.

type OrientedDomino a = Oriented (Domino a)

A type alias that specifies that OrientedDomino is an Oriented type that wraps a Domino item.

type SimpleDomino = Domino DieValue

A SimpleDomino is a Domino that contains DieValue objects, it thus can not have an "empty" value.

type ComplexDomino = Domino (Maybe DieValue)

A ComplexDomino is a Domino that contains Maybe values wrapping a DieValue. In case of a Nothing, that side is considered empty.

Render domino values

dominoH

Arguments

:: ComplexDomino
The ComplexDomino object to render horizontally.

-> Char
The unicode character that represents the given ComplexDomino value in a horizontal manner.

Convert a ComplexDomino value to a unicode character rendering the domino value horizontally.

dominoH'

Arguments

:: SimpleDomino
The SimpleDomino object to render horizontally.

-> Char
The unicode character that represents the given SimpleDomino value in a horizontal manner.

Convert a SimpleDomino value to a unicode character rendering the domino value horizontally.

dominoV

Arguments

:: ComplexDomino
The ComplexDomino object to render vertically.

-> Char
The unicode character that represents the given ComplexDomino value in a vertical manner.

Convert a ComplexDomino value to a unicode character rendering the domino value vertically.

dominoV'

Arguments

:: SimpleDomino
The SimpleDomino object to render vertically.

-> Char
The unicode character that represents the given SimpleDomino value in vertical manner.

Convert a SimpleDomino value to a unicode character rendering the domino value vertically.

domino

Arguments

:: OrientedDomino (Maybe DieValue)
The OrientedDomino to render.

-> Char
The unicode characters that represents the OrientedDomino value.

Convert an OrientedDomino to its unicode equivalent, where the sides of the domino can be empty.

domino'

Arguments

:: OrientedDomino DieValue
The OrientedDomino to render.

-> Char
The unicode characters that represents the OrientedDomino value.

Convert an OrientedDomino to its unicode equivalent, where the sides of the domino can not be empty.

Convert from Characters

fromDomino

Arguments

:: Char
The given Character to convert to an Oriented ComplexDomino object.

-> Maybe (Oriented ComplexDomino)
The equivalent Oriented ComplexDomino object for the given Character wrapped in a Just; Nothing if the character is not a domino character.

Convert the given Character to an Oriented ComplexDomino object. If the given Character wrapped in a Just data constructor if the Character is a valid domino character; otherwise Nothing.

fromDomino'

Arguments

:: Char
The given Character to convert to an Oriented ComplexDomino object.

-> Oriented ComplexDomino
The equivalent Oriented ComplexDomino object for the given Character.

Convert the given Character to an Oriented ComplexDomino object. If the given Character is not a valid domino character, the result is unspecified.