Data.Char.Emoji

Description

Unicode defines 2182 emoji characters, this module aims to make working with emoji characters more convenient.

Clock emoji

data Clock

A Clock object that can be converted to a unicode character that displays a clock with the given time. The Clock has an hours field that contains the given hours between 0 and 12, and a minutes30 field that if True, means that the clock is half past that hour.

hours :: Clock -> Int

The number of hours on the given clock. Is between 0 and 12. For 0, the minutes30 is True; and for 12, the minutes30 is False.

minutes30 :: Clock -> Bool

Is True if it is half past the given hour on the Clock.

clock

Arguments

:: Int
The given hour of the clock, can be any value, but will be set between 1 and 12.

-> Bool
A Boolean that indicates if it is half past that hour, so True means we add 30 minutes.

-> Clock
A clock object that represents the time that is passed through an hour and .

Construct a Clock object with the given number of hours, and a Boolean that indicates if it is half past that hour. The function will ensure that the hours are between 0 and 12 (both inclusive). For half past 12, we use half past 0, for 12 hours, we use simply 12.

closestClock

Arguments

:: Int
The number of hours.

-> Int
The number of minutes, must be between 0 and 60.

-> Clock
The clock object that is the closest to the given hours and minutes.

Generate the Clock object that is the closest to the given hours and minutes.

Skin color modifier

data SkinColorModifier

Some emoji deal with people. One can change the color of the skin with the SkinColorModifier. For the skin color, the /Fitzpatrick scale/ is used. A numerical classification system for skin types.

Constructors

Light
An emoji modifier that applies Fitzpatrick skin type one or two to the Emoji.

MediumLight
An emoji modifier that applies Fitzpatrick skin type three to the Emoji.

Medium
An emoji modifier that applies Fitzpatrick skin type four to the Emoji.

MediumDark
An emoji modifier that applies Fitzpatrick skin type five to the Emoji.

Dark
An emoji modifier that applies Fitzpatrick skin type six to the Emoji.

fromFitzPatrick

Arguments

:: Integral i

=> i
The given Fitzpatrick skin type.

-> Maybe SkinColorModifier
The corresponding SkinColorModifier wrapped in a Just; Nothing if no such modifier exists.

Convert the given Fitzpatrick skin type to the corresponding SkinColorModifier wrapped in a Just, if no such SkinColorModifier exists, Nothing is returned.

Pattern synonyms for the SkinColorModifier elements

pattern FitzPatrickI :: SkinColorModifier

The SkinColorModifier that corresponds to type one of the /Fitzpatrick scale/.

pattern FitzPatrickII :: SkinColorModifier

The SkinColorModifier that corresponds to type two of the /Fitzpatrick scale/.

pattern FitzPatrickIII :: SkinColorModifier

The SkinColorModifier that corresponds to type three of the /Fitzpatrick scale/.

pattern FitzPatrickIV :: SkinColorModifier

The SkinColorModifier that corresponds to type four of the /Fitzpatrick scale/.

pattern FitzPatrickV :: SkinColorModifier

The SkinColorModifier that corresponds to type five of the /Fitzpatrick scale/.

pattern FitzPatrickVI :: SkinColorModifier

The SkinColorModifier that corresponds to type six of the /Fitzpatrick scale/.

Submodule import

module Data.Char.Emoji.Core

module Data.Char.Emoji.BloodType

module Data.Char.Emoji.Flag

module Data.Char.Emoji.Gender

module Data.Char.Emoji.Moon

module Data.Char.Emoji.Zodiac