Data.Char.Emoji.BloodType

Description

The Unicode standard defines four emoji for the O, B, A, and AB blood type.

Blood type emoji

data BloodType

A BloodType object used to convert to its unicode equivalent. The BloodType is also seen as a 2-bit value with the leftmost bit representing the presence of A antigens and the rightmost the presence of B antigens.

Constructors

O
The O blood type, with no presence of A and B antigens, denoted by 🅾️.

B
The B blood type, with presence of the B antigen, denoted by 🅱️.

A
The A blood type, with presence of the A antigen, denoted by 🅰️.

AB
The AB blood type, with presence of the A and B antigens, denoted by 🆎.

Drop of blood emoji

pattern DropOfBlood :: Char

A emoji that depicts a drop of blood. This looks like 🩸.