Data.Char.Emoji.Clock

Description

The emoji have 24 clock emoji, each time with 30 minutes difference.

Clock emoji

data Clock

A Clock object that can be converted to a unicode character that displays a clock with the given time. The Clock has an hours field that contains the given hours between 0 and 12, and a minutes30 field that if True, means that the clock is half past that hour.

hours :: Clock -> Int

The number of hours on the given clock. Is between 0 and 12. For 0, the minutes30 is True; and for 12, the minutes30 is False.

minutes30 :: Clock -> Bool

Is True if it is half past the given hour on the Clock.

clock

Arguments

:: Int
The given hour of the clock, can be any value, but will be set between 1 and 12.

-> Bool
A Boolean that indicates if it is half past that hour, so True means we add 30 minutes.

-> Clock
A clock object that represents the time that is passed through an hour and .

Construct a Clock object with the given number of hours, and a Boolean that indicates if it is half past that hour. The function will ensure that the hours are between 0 and 12 (both inclusive). For half past 12, we use half past 0, for 12 hours, we use simply 12.

closestClock

Arguments

:: Int
The number of hours.

-> Int
The number of minutes, must be between 0 and 60.

-> Clock
The clock object that is the closest to the given hours and minutes.

Generate the Clock object that is the closest to the given hours and minutes.