Data.Char.Emoji.Hand

Description

Unicode has emoji's for hands. In this module we make it more convenient to render hand gestures with a specific skin color.

Documentation

data SingleCharHandGesture

A datatype that constructs hand gestures that correspond with a single Character.

Constructors

WavingHand
A waving hand, this is denoted with 👋.

RaisedBackOfHand
The raised back of a hand, this is denoted with 🤚.

RaisedHand
A raised hand, this is denoted with ✋.

VulcanSalute
The Vulcan salute, this is denoted with 🖖.

OkHandSign
The okay hand sign, this is denoted with 👌.

PinchedFingers
The pinched fingers gesture, this is denoted with 🤌.

PinchingHand
The pinching hand gesture, this is denoted with 🤏.

CrossedFingers
The crossed fingers gesture, this is denoted with 🤞.

LoveYouGesture
The love you gesture, this is denoted with 🤟.

SignOfTheHorns
The sign of the horns, this is denoted with 🤘.

CallMeHand
The call me hand sign, this is denoted with 🤙.

data MultiCharHandGesture

Emoji with hands that map on a sequence of characters instead of one character.

pattern FingersCrossed :: SingleCharHandGesture

A pattern synonym for CrossedFingers.

pattern SpockHand :: SingleCharHandGesture

A pattern synonym for the VulcanSalute.

pattern HornsSign :: SingleCharHandGesture

A pattern synonym for SignOfTheHorns.