Data.Char.Emoji.NoEvilMonkey

Description

Unicode has three codepoints that refer to characters of monkeys where the first one does not see, the second does not hear, and the third one does not speak.

Documentation

data NoEvilMonkey

A data constructors that lists the three different no-evil monkeys.

Constructors

SeeNoEvilMonkey
The see-no-evil monkey, denoted with 🙈.

HearNoEvilMonkey
The hear-no-evil monkey, denoted with 🙉.

SpeakNoEvilMonkey
The speak-no-evil monkey, denoted with 🙊.