Data.Char.Emoji.Science

Description

Unicode defines nine Emoji related to science. This module has a data type together with functions to convert this to and from a text fragment.

Defining science emoji.

data ScienceEmoji

There are nine emoji that depict science.

Constructors

Alembic

TestTube
A test tube is used to conduct chemical experiments. Normally this is depicted as 🧪.

PetriDish
A petri dish is used to culture microbes. Normally this is depicted as 🧫.

DnaDoubleHelix
A double helix of DNA is the genetic blueprint of life. Normally this is depicted as 🧬.

Microscope
A microscope is used to magnify small objects. Normally this is depicted as 🔬.

Telescope
A telescope is used to gaze at stars and planets in the night sky. Normally this is depicted as 🔭.

SatelliteAntenna
A dish satellite is used to send and receive to or from communication satellites. This is normally depicited as 📡.