Data.Char.Emoji.SkinColor

Description

For several emoji, one can specify the color of the skin of the person(s) of the emoji. This module defines the skin color modifiers together with its values on the Fitzpatrick scale.

Skin color modifier

data SkinColorModifier

Some emoji deal with people. One can change the color of the skin with the SkinColorModifier. For the skin color, the /Fitzpatrick scale/ is used. A numerical classification system for skin types.

Constructors

Light
An emoji modifier that applies Fitzpatrick skin type one or two to the Emoji.

MediumLight
An emoji modifier that applies Fitzpatrick skin type three to the Emoji.

Medium
An emoji modifier that applies Fitzpatrick skin type four to the Emoji.

MediumDark
An emoji modifier that applies Fitzpatrick skin type five to the Emoji.

Dark
An emoji modifier that applies Fitzpatrick skin type six to the Emoji.

type OptionalSkinColorModifier = Maybe SkinColorModifier

For emoji often the skin color is optional: in case there is no skin color specified, the emoji have often a yellow skin color.

fromFitzpatrick

Arguments

:: Integral i

=> i
The given Fitzpatrick skin type.

-> Maybe SkinColorModifier
The corresponding SkinColorModifier wrapped in a Just; Nothing if no such modifier exists.

Convert the given Fitzpatrick skin type to the corresponding SkinColorModifier wrapped in a Just, if no such SkinColorModifier exists, Nothing is returned.

isSkinColorModifier

Arguments

:: Char
The given Character to check.

-> Bool
True if the given Character is a skin color modifier, False otherwise.

Check if the given Character is a skin color modifier.

Create emoji with a SkinColorModifier

class UnicodeText a => WithSkinColorModifierUnicodeText a where

A typeclass where one can specify that the object can be rendered with a given skin color modifier.

Minimal complete definition

Nothing

Methods

withSkinModifier

Arguments

:: a
The given item to render to a unicode Text object.

-> SkinColorModifier
The given skin color modifier to apply.

-> Text
The corresponding Text where we applied the given SkinColorModifier.

Apply the given SkinColorModifier to the item and obtain a Text object where the item has been modified with the SkinColorModifier.

withOptionalSkinModifier

Arguments

:: a
The given item to render to a unicode Text object.

-> OptionalSkinColorModifier
The given optional skin color modifier.

-> Text
The corresponding Text where we applied the given SkinColorModifier.

Apply the given SkinColorModifier to the item given it is not Nothing such that the object is rendered with the given skin color modifier.

withoutOptionalSkinModifier

Arguments

:: Text
The given Text object that should be decoded.

-> Maybe (a, OptionalSkinColorModifier)
An optional 2-tuple with the item that has been read, and an optional SkinColorModifier.

Convert the given Text to an item with an OptionalSkinColorModifier that might have been applied.

withSkinModifier'

Arguments

:: Text
The given Text object where we want to specify the skin color.

-> SkinColorModifier
The givenSkinColorModifier to apply.

-> Text
The given Text object combined with the given SkinColorModifier.

Append the given Text object with the Unicode character to modify its skin color.

withOptionalSkinModifier'

Arguments

:: Text
The given Text object where we want to specify the skin color.

-> OptionalSkinColorModifier
The givenOptionalSkinColorModifier to apply.

-> Text
The given Text object combined with the given SkinColorModifier.

Append the given Text object with the Unicode character to modify its skin color. If Nothing, then no modification is applied.

withoutOptionalSkinModifier'

Arguments

:: Text
The given Text to decompose.

-> (Text, OptionalSkinColorModifier)
A 2-tuple where the first item is the remaining Text and where the second item is an optioanl SkinColorModifier.

Convert the given Text object to a wrapped Text object with an OptionalSkinColorModifier.

Pattern synonyms for the SkinColorModifier elements

pattern FitzpatrickI :: SkinColorModifier

The SkinColorModifier that corresponds to type one of the /Fitzpatrick scale/.

pattern FitzpatrickII :: SkinColorModifier

The SkinColorModifier that corresponds to type two of the /Fitzpatrick scale/.

pattern FitzpatrickIII :: SkinColorModifier

The SkinColorModifier that corresponds to type three of the /Fitzpatrick scale/.

pattern FitzpatrickIV :: SkinColorModifier

The SkinColorModifier that corresponds to type four of the /Fitzpatrick scale/.

pattern FitzpatrickV :: SkinColorModifier

The SkinColorModifier that corresponds to type five of the /Fitzpatrick scale/.

pattern FitzpatrickVI :: SkinColorModifier

The SkinColorModifier that corresponds to type six of the /Fitzpatrick scale/.