Data.Char.Math

Description

Introduction

This module aims to make it more convenient to transform mathematical characters to their double struck, fraktur, calligraphic, etc. equivalent.

Most of the characters are defined in the 1D400–1D7FF Unicode block Mathematical Alphanumeric Symbols. See also the Wikipedia page.

Examples

>>> math Serif Italic Bold 'x'
Just '\119961'
>>> math Serif Italic Bold '3'
Just '\120785'
>>> latinMath Serif Italic Bold 'x'
Just '\119961'
>>> latinMath Serif Italic Bold '3'
Nothing
>>> script NoBold 'S'
Just '\119982'
>>> intToDigitChar SansSerif Bold 3
Just '\120815'
>>> intToDigitChar SansSerif Bold 33
Nothing

Supported ranges of characters

The transformations of this module only supports the following small subset of Unicode points:

ASCII latin letters

A–Z and a–z ranges

Greek-like symbols

The following characters from the Unicode block Greek and Coptic: ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩϴαβγδεζηθικλμνξοπρςστυφχψωϵϑϰϕϱϖ.

The following characters from the Unicode block Mathematical Operators: ∇∂.

Digits

0–9 range

Naming Conventions

The functions with a name finishing with a quote ' (such as math') do not check their input: they transform the characters in the supported range and have an unspecified behaviour outside this range.

The functions without a quote in their name (such as math) check their input and output and wrap the resulting transformation with Maybe.

Serif/sans-serif mathematical symbols

math

Arguments

:: FontStyle
The given FontStyle to use.

-> ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with the given font style, in the given emphasis and in the given italics type wrapped in a Just if the character is supported (see: supported ranges) If the character is outside theses ranges, Nothing is returned.

math'

Arguments

:: FontStyle
The given FontStyle to use.

-> ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted in the given FontStyle, depending on the given Emphasis in bold or not, and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with the given font style, with a given emphasis and a given italics style. This maps characters an equivalent sans-serif symbol for characters in the supported ranges. For characters outside the range, the behavior is unspecified.

Latin-only characters

latinMath

Arguments

:: FontStyle
The given FontStyle to use.

-> ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with the given font style, in the given emphasis and in the given italics type wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

mathAlpha

Arguments

:: FontStyle
The given FontStyle to use.

-> ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Deprecated: Use latinMath

Convert the given character to a mathematical symbol with the given font style, in the given emphasis and in the given italics type wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

mathAlpha'

Arguments

:: FontStyle
The given FontStyle to use.

-> ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted in the given FontStyle, depending on the given Emphasis in bold or not, and depending on the given ItalicType in italics or not.

Deprecated: Use math'

Convert the given character to a mathematical symbol with the given font style, with a given emphasis and a given italics style. This maps characters an equivalent sansSerif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

Serif mathematical symbols

serif

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and in the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serif'

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and a given italics style. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifNoBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold and in the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifNoBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, not in bold and in a italics type. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold with the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, in bold and in a italics type. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifNoItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in the given emphasis and not in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifNoItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and not in italics.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and not in italics. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in the given emphasis and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and in italics.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and in italics. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifNoBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold, and no italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifNoBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and not in italics.

Convert the given character to a mathematical symbol with serifs, with no bold and no italics. This maps characters to itself for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold, and no italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and not in italics.

Convert the given character to a mathematical symbol with serifs, in bold not in italics. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifNoBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold, and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifNoBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and in italics.

Convert the given character to a mathematical symbol with serifs, with no bold and in italics. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold, and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and in italics.

Convert the given character to a mathematical symbol with serifs, with in bold and in italics. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

Sans-serif mathematical symbols

sansSerif

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and in the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerif'

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and a given italics style. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifNoBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold and in the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifNoBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, not in bold and in a italics type. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold with the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, in bold and in a italics type. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifNoItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and not in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifNoItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and not in italics.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and not in italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and in italics.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and in italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifNoBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold, and no italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifNoBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and not in italics.

Convert the given character to a mathematical symbol without serifs, with no bold and no italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold, and no italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and not in italics.

Convert the given character to a mathematical symbol without serifs, in bold not in italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifNoBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold, and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifNoBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and in italics.

Convert the given character to a mathematical symbol without serifs, with no bold and in italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold, and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and in italics.

Convert the given character to a mathematical symbol without serifs, with in bold and in italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

Digit characters

Char-based conversion

digit

Arguments

:: FontStyle
The given font style.

-> Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Maybe Char
The corresponding symbol in serifs for the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character with the given Emphasis in the given font style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digit'

Arguments

:: FontStyle
The given font style.

-> Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Char
The corresponding symbol in the given font style for the given emphasis style, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character with a given Emphasis in the given font style. The result for characters outside this range is unspecified.

digitSansSerif

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Maybe Char
The corresponding symbol in sans-serifs for the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character with the given Emphasis in sans-serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSansSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Char
The corresponding symbol in sans-serifs for the given emphasis style, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character with a given Emphasis in sans-serif style. The result for characters outside this range is unspecified.

digitSerif

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Maybe Char
The corresponding symbol in serifs for the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character with the given Emphasis in serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Char
The corresponding symbol in serifs for the given emphasis style, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character with a given Emphasis in serif style. The result for characters outside this range is unspecified.

digitSerifRegular

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSerifRegular'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in serifs not in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold serif style. The result for characters outside this range is unspecified.

digitSerifBold

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a bold serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSerifBold'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in serifs in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a bold serif style. The result for characters outside this range is unspecified.

digitSansSerifRegular

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in sans-serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold sans-serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSansSerifRegular'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in sans-serifs not in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold sans-serif style. The result for characters outside this range is unspecified.

digitSansSerifBold

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in sans-serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a bold sans-serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSansSerifBold'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in sans-serifs in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a bold sans-serif style. The result for characters outside this range is unspecified.

digitMonospace

Arguments

:: Char
The given digit character to convert.

-> Maybe Char
The corresponding symbol in monospace style wrapped in a Just, Nothing if the character is outside the range.

Converts the given digit (0–9) charcters to its equivalent in monospace style wrapped in a Just data constructor. Nothing for characters outside the range.

digitMonospace'

Arguments

:: Char
The given digit character to convert.

-> Char
The corresponding character in monspace style. Unspecified outside the digit range.

Converts the given digit (0–9) charcters to its equivalent in monospace style. Unspecified result for characters outside the range.

digitDoubleStruck

Arguments

:: Char
The given digit character to convert.

-> Maybe Char
The corresponding symbol in double-struck style wrapped in a Just, Nothing if the character is outside the range.

Converts the given digit (0–9) charcters to its equivalent in double-struck style wrapped in a Just data constructor. Nothing for characters outside the range.

digitDoubleStruck'

Arguments

:: Char
The given digit character to convert.

-> Char
The corresponding character in double-struck style. Unspecified outside the digit range.

Converts the given digit (0–9) charcters to its equivalent in double-struck style. Unspecified result for characters outside the range.

Int to digit characters

intToDigitChar

Arguments

:: FontStyle
The given font style.

-> Emphasis
The given emphasis style.

-> Int
The given number to convert.

-> Maybe Char
The corresponding symbol in the given font style in the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character with the given Emphasis in the given FontStyle wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitChar'

Arguments

:: FontStyle
The given font style.

-> Emphasis
The given emphasis style.

-> Int
The given number to convert.

-> Char
The corresponding symbol in sans-serifs in the given font style the given emphasis style, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character with a given Emphasis in the given FontStyle. The result for numbers outside this range is unspecified.

intToDigitSerif

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert

-> Maybe Char
The corresponding symbol in serifs in the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character with the given Emphasis in serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert.

-> Char
The corresponding symbol in serifs in the given emphasis style, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character with a given Emphasis in serif style. The result for numbers outside this range is unspecified.

intToDigitSansSerif

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert

-> Maybe Char
The corresponding symbol in sans-serifs in the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character with the given Emphasis in sans-serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSansSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert.

-> Char
The corresponding symbol in sans-serifs in the given emphasis style, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character with a given Emphasis in sans-serif style. The result for numbers outside this range is unspecified.

intToDigitSerifRegular

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in a non-bold serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSerifRegular'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding symbol in serifs not in bold, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character in a non-bold serif style. The result for numbers outside this range is unspecified.

intToDigitSerifBold

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in a bold serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSerifBold'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding symbol in serifs in bold, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character in a bold serif style. The result for numbers outside this range is unspecified.

intToDigitMonospace

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in monospace style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in monospace style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitMonospace'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding character in monspace style. Unspecified outside the digit range.

Convert the given number (0–9) to its corresponding character in monospace style. Unspecified result for numbers outside this range.

intToDigitDoubleStruck

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in monospace style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in double-struck style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitDoubleStruck'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding character in double-struck style. Unspecified outside the digit range.

Convert the given number (0–9) to its corresponding character in double-struck style. Unspecified result for numbers outside this range.

Monospace symbols

monospace

Arguments

:: Char
The given character to convert to a monspace symbol.

-> Maybe Char
The equivalent monospace character wrapped in a Just data constructor, Nothing if outside the alphanumerical range.

Convert the given character to its monospace equivalent for the alphabet and numerical range (A–Z, a–z, and 0–9) wrapped in a Just data constructor. For characters outside the range, Nothing is returned.

monospace'

Arguments

:: Char
The given character to convert to a monospace symbol.

-> Char
The equivalent monospace symbol for the given character.

Convert the given character to its monospace equivalent for the alphabet and numerical range (A–Z, a–z, and 0–9). For characters outside the range, the result is unspecified.

Double struck symbols

doubleStruck

Arguments

:: Char
The character to convert to a double struck symbol.

-> Maybe Char
The double struck symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent double stuck character.

Obtain the double struck symbol for the given character. The supported range of characters are the alphabet characters (A–Z, and a–z), and the numerical characters (0–9). The symbols are wrapped in the Just data constructor. For characters outside the range, Nothing is returned.

doubleStruck'

Arguments

:: Char
The character to convert to a double struck symbol.

-> Char
The double struck symbol for the given character. If the character is not an ASCII alphanumerical character, the result is unspecified.

Obtain the double struck symbol for the given character. The supported range of characters are the alphabet character (A–Z, and a–z), and the numerical characters (0–9). For characters other than these, the behaviour is unspecified.

Script (or calligraphic symbols)

script

Arguments

:: Emphasis
The given Emphasis style to use.

-> Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is not written in the given Emphasis style. If the character is not supported (see: Data.Char.Math), Nothing is returned.

script'

Arguments

:: Emphasis
The given Emphasis style to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, and depending on the Emphasis in bold or not.

Convert the given character to its script or calligraphic symbol. This symbol is written in the given Emphasis style. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

scriptRegular

Arguments

:: Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is not written in boldface. If the character is not supported (see: Data.Char.Math), Nothing is returned.

scriptRegular'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, not in bold.

Convert the given character to its script or calligraphic symbol. This symbol is not written in boldface. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

scriptBold

Arguments

:: Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is written in boldface. If the character is not supported (see: Data.Char.Math), Nothing is returned.

scriptBold'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, and in bold.

Convert the given character to its script or calligraphic symbol. This symbol is written in boldface. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

calligraphy

Arguments

:: Emphasis
The given Emphasis style to use.

-> Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is not written in the given Emphasis style. If the character is not supported (see: Data.Char.Math), Nothing is returned. This is an alias of script.

calligraphy'

Arguments

:: Emphasis
The given Emphasis style to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, and depending on the Emphasis in bold or not.

Convert the given character to its script or calligraphic symbol. This symbol is written in the given Emphasis style. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified. This is an alias of script'.

calligraphyRegular

Arguments

:: Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is not written in boldface. If the character is not supported (see: Data.Char.Math), Nothing is returned. This is an alias of scriptRegular.

calligraphyRegular'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, not in bold.

Convert the given character to its script or calligraphic symbol. This symbol is not written in boldface. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified. This is an alias of scriptRegular'.

calligraphyBold

Arguments

:: Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is written in boldface. If the character is not supported (see: Data.Char.Math), Nothing is returned. This is an alias of scriptBold.

calligraphyBold'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, and in bold.

Convert the given character to its script or calligraphic symbol. This symbol is written in boldface. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified. This is an alias of scriptBold'.

Fraktur symbols

fraktur

Arguments

:: Emphasis
The given emphasis style to use.

-> Char
The character to convert to a fraktur symbol with the given emphasis style.

-> Maybe Char
The fraktur symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent fraktur character.

Obtain the fraktur symbol for the given character in the given emphais style. The result is wrapped in a Just data constructor. The range of supported characters are the alphabet characters (A–Z, and a–z). In case a character outside the range is passed to the function, Nothing is returned.

fraktur'

Arguments

:: Emphasis
The given emphasis style to use.

-> Char
The character to convert to a fraktur symbol in the given emphasis style.

-> Char
The equivalent fraktur charater for the given character.

Obtain the fraktur symbol for the given character in the given emphasis style. The supported range of characters are the alphabet characters (A–Z, and a–z). In case the character is not in this range, it is unspecified what will be returned.

frakturRegular

Arguments

:: Char
The character to convert to a regular fraktur symbol.

-> Maybe Char
The fraktur symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent fraktur character.

Obtain the fraktur symbol for the given character in a regular (not bold style). The result is wrapped in a Just data constructor. The range of supported characters are the alphabet characters (A–Z, and a–z). In case a character outside the range is passed to the function, Nothing is returned.

frakturRegular'

Arguments

:: Char
The character to convert to a regular fraktur symbol.

-> Char
The equivalent fraktur charater for the given character.

Obtain the fraktur symbol for the given character in a regular (not bold) style. The supported range of characters are the alphabet characters (A–Z, and a–z). In case the character is not in this range, it is unspecified what will be returned.

frakturBold

Arguments

:: Char
The character to convert to a bold fraktur symbol.

-> Maybe Char
The fraktur symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent fraktur character.

Obtain the fraktur symbol for the given character in a bold. The result is wrapped in a Just data constructor. The range of supported characters are the alphabet characters (A–Z, and a–z). In case a character outside the range is passed to the function, Nothing is returned.

frakturBold'

Arguments

:: Char
The character to convert to a bold fraktur symbol.

-> Char
The equivalent fraktur charater for the given character.

Obtain the fraktur symbol for the given character in a bold style. The supported range of characters are the alphabet characters (A–Z, and a–z). In case the character is not in this range, it is unspecified what will be returned.