Data.Char.Math.DoubleStruck

Description

See Data.Char.Math for further documentation.

Letters

doubleStruck

Arguments

:: Char
The character to convert to a double struck symbol.

-> Maybe Char
The double struck symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent double stuck character.

Obtain the double struck symbol for the given character. The supported range of characters are the alphabet characters (A–Z, and a–z), and the numerical characters (0–9). The symbols are wrapped in the Just data constructor. For characters outside the range, Nothing is returned.

doubleStruck'

Arguments

:: Char
The character to convert to a double struck symbol.

-> Char
The double struck symbol for the given character. If the character is not an ASCII alphanumerical character, the result is unspecified.

Obtain the double struck symbol for the given character. The supported range of characters are the alphabet character (A–Z, and a–z), and the numerical characters (0–9). For characters other than these, the behaviour is unspecified.

Digits

digitDoubleStruck

Arguments

:: Char
The given digit character to convert.

-> Maybe Char
The corresponding symbol in double-struck style wrapped in a Just, Nothing if the character is outside the range.

Converts the given digit (0–9) charcters to its equivalent in double-struck style wrapped in a Just data constructor. Nothing for characters outside the range.

digitDoubleStruck'

Arguments

:: Char
The given digit character to convert.

-> Char
The corresponding character in double-struck style. Unspecified outside the digit range.

Converts the given digit (0–9) charcters to its equivalent in double-struck style. Unspecified result for characters outside the range.

intToDigitDoubleStruck

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in monospace style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in double-struck style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitDoubleStruck'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding character in double-struck style. Unspecified outside the digit range.

Convert the given number (0–9) to its corresponding character in double-struck style. Unspecified result for numbers outside this range.