Data.Char.Math.Fraktur

Description

See Data.Char.Math for further documentation.

Documentation

fraktur

Arguments

:: Emphasis
The given emphasis style to use.

-> Char
The character to convert to a fraktur symbol with the given emphasis style.

-> Maybe Char
The fraktur symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent fraktur character.

Obtain the fraktur symbol for the given character in the given emphais style. The result is wrapped in a Just data constructor. The range of supported characters are the alphabet characters (A–Z, and a–z). In case a character outside the range is passed to the function, Nothing is returned.

fraktur'

Arguments

:: Emphasis
The given emphasis style to use.

-> Char
The character to convert to a fraktur symbol in the given emphasis style.

-> Char
The equivalent fraktur charater for the given character.

Obtain the fraktur symbol for the given character in the given emphasis style. The supported range of characters are the alphabet characters (A–Z, and a–z). In case the character is not in this range, it is unspecified what will be returned.

frakturRegular

Arguments

:: Char
The character to convert to a regular fraktur symbol.

-> Maybe Char
The fraktur symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent fraktur character.

Obtain the fraktur symbol for the given character in a regular (not bold style). The result is wrapped in a Just data constructor. The range of supported characters are the alphabet characters (A–Z, and a–z). In case a character outside the range is passed to the function, Nothing is returned.

frakturRegular'

Arguments

:: Char
The character to convert to a regular fraktur symbol.

-> Char
The equivalent fraktur charater for the given character.

Obtain the fraktur symbol for the given character in a regular (not bold) style. The supported range of characters are the alphabet characters (A–Z, and a–z). In case the character is not in this range, it is unspecified what will be returned.

frakturBold

Arguments

:: Char
The character to convert to a bold fraktur symbol.

-> Maybe Char
The fraktur symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent fraktur character.

Obtain the fraktur symbol for the given character in a bold. The result is wrapped in a Just data constructor. The range of supported characters are the alphabet characters (A–Z, and a–z). In case a character outside the range is passed to the function, Nothing is returned.

frakturBold'

Arguments

:: Char
The character to convert to a bold fraktur symbol.

-> Char
The equivalent fraktur charater for the given character.

Obtain the fraktur symbol for the given character in a bold style. The supported range of characters are the alphabet characters (A–Z, and a–z). In case the character is not in this range, it is unspecified what will be returned.