Data.Char.Math.Monospace

Description

See Data.Char.Math for further documentation.

Letters

monospace

Arguments

:: Char
The given character to convert to a monspace symbol.

-> Maybe Char
The equivalent monospace character wrapped in a Just data constructor, Nothing if outside the alphanumerical range.

Convert the given character to its monospace equivalent for the alphabet and numerical range (A–Z, a–z, and 0–9) wrapped in a Just data constructor. For characters outside the range, Nothing is returned.

monospace'

Arguments

:: Char
The given character to convert to a monospace symbol.

-> Char
The equivalent monospace symbol for the given character.

Convert the given character to its monospace equivalent for the alphabet and numerical range (A–Z, a–z, and 0–9). For characters outside the range, the result is unspecified.

Digits

digitMonospace

Arguments

:: Char
The given digit character to convert.

-> Maybe Char
The corresponding symbol in monospace style wrapped in a Just, Nothing if the character is outside the range.

Converts the given digit (0–9) charcters to its equivalent in monospace style wrapped in a Just data constructor. Nothing for characters outside the range.

digitMonospace'

Arguments

:: Char
The given digit character to convert.

-> Char
The corresponding character in monspace style. Unspecified outside the digit range.

Converts the given digit (0–9) charcters to its equivalent in monospace style. Unspecified result for characters outside the range.

intToDigitMonospace

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in monospace style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in monospace style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitMonospace'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding character in monspace style. Unspecified outside the digit range.

Convert the given number (0–9) to its corresponding character in monospace style. Unspecified result for numbers outside this range.