Data.Char.Math.SansSerif

Description

See Data.Char.Math for further documentation.

Sans-serif mathematical alphabet symbols

sansSerif

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and in the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerif'

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and a given italics style. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifNoBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold and in the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifNoBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, not in bold and in a italics type. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold with the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, in bold and in a italics type. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifNoItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and not in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifNoItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and not in italics.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and not in italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and in italics.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and in italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifNoBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold, and no italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifNoBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and not in italics.

Convert the given character to a mathematical symbol without serifs, with no bold and no italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold, and no italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and not in italics.

Convert the given character to a mathematical symbol without serifs, in bold not in italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifNoBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold, and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifNoBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and in italics.

Convert the given character to a mathematical symbol without serifs, with no bold and in italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

sansSerifBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold, and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

sansSerifBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and in italics.

Convert the given character to a mathematical symbol without serifs, with in bold and in italics. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

Digit characters

Character conversion

digitSansSerif

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Maybe Char
The corresponding symbol in sans-serifs for the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character with the given Emphasis in sans-serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSansSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Char
The corresponding symbol in sans-serifs for the given emphasis style, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character with a given Emphasis in sans-serif style. The result for characters outside this range is unspecified.

digitSansSerifRegular

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in sans-serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold sans-serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSansSerifRegular'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in sans-serifs not in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold sans-serif style. The result for characters outside this range is unspecified.

digitSansSerifBold

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in sans-serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a bold sans-serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSansSerifBold'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in sans-serifs in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a bold sans-serif style. The result for characters outside this range is unspecified.

Int to digit characters

intToDigitSansSerif

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert

-> Maybe Char
The corresponding symbol in sans-serifs in the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character with the given Emphasis in sans-serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSansSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert.

-> Char
The corresponding symbol in sans-serifs in the given emphasis style, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character with a given Emphasis in sans-serif style. The result for numbers outside this range is unspecified.

intToDigitSansSerifRegular

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in sans-serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in a non-bold sans-serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSansSerifRegular'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding symbol in sans-serifs not in bold, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character in a non-bold sans-serif style. The result for numbers outside this range is unspecified.

intToDigitSansSerifBold

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in sans-serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in a bold sans-serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSansSerifBold'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding symbol in sans-serifs in bold, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character in a bold sans-serif style. The result for numbers outside this range is unspecified.