Data.Char.Math.SansSerif.Digit

Description

See Data.Char.Math for further documentation.

Character conversion

digitSansSerif

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Maybe Char
The corresponding symbol in sans-serifs for the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character with the given Emphasis in sans-serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSansSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Char
The corresponding symbol in sans-serifs for the given emphasis style, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character with a given Emphasis in sans-serif style. The result for characters outside this range is unspecified.

digitSansSerifRegular

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in sans-serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold sans-serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSansSerifRegular'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in sans-serifs not in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold sans-serif style. The result for characters outside this range is unspecified.

digitSansSerifBold

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in sans-serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a bold sans-serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSansSerifBold'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in sans-serifs in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a bold sans-serif style. The result for characters outside this range is unspecified.

Int to digit characters

intToDigitSansSerif

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert

-> Maybe Char
The corresponding symbol in sans-serifs in the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character with the given Emphasis in sans-serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSansSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert.

-> Char
The corresponding symbol in sans-serifs in the given emphasis style, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character with a given Emphasis in sans-serif style. The result for numbers outside this range is unspecified.

intToDigitSansSerifRegular

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in sans-serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in a non-bold sans-serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSansSerifRegular'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding symbol in sans-serifs not in bold, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character in a non-bold sans-serif style. The result for numbers outside this range is unspecified.

intToDigitSansSerifBold

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in sans-serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in a bold sans-serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSansSerifBold'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding symbol in sans-serifs in bold, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character in a bold sans-serif style. The result for numbers outside this range is unspecified.