Data.Char.Math.SansSerif.Greek

Description

See Data.Char.Math for further documentation.

Documentation

greekSansSerif

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and in the given italics type wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSansSerif'

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and a given italics style. This maps characters an equivalent sans-serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSansSerifNoBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold and in the given italics type wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSansSerifNoBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, not in bold and in a italics type. This maps characters an equivalent sans-serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSansSerifBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold with the given italics type wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSansSerifBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, in bold and in a italics type. This maps characters an equivalent sans-serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSansSerifNoItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and not in italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSansSerifNoItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and not in italics.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and not in italics. This maps characters an equivalent sans-serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSansSerifItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and in italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSansSerifItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and in italics.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and in italics. This maps characters an equivalent sans-serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSansSerifNoBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold and no italics. This maps characters to itself wrapped in a Just; Nothing if that character does not exists.

greekSansSerifNoBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and not in italics.

Convert the given character to a mathematical symbol without serifs, with no bold and no italics. This maps characters to itself.

greekSansSerifBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold, and no italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSansSerifBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and not in italics.

Convert the given character to a mathematical symbol without serifs, in bold not in italics. This maps characters an equivalent sans-serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSansSerifNoBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold, and in italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSansSerifNoBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and in italics.

Convert the given character to a mathematical symbol without serifs, with no bold and no italics. This maps characters to itself.

greekSansSerifBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold, and in italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSansSerifBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and in italics.

Convert the given character to a mathematical symbol without serifs, with in bold and in italics. This maps characters an equivalent sans-serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.