Data.Char.Math.SansSerif.Latin

Description

See Data.Char.Math for further documentation.

Documentation

latinSansSerif

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and in the given italics type wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSansSerif'

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and a given italics style. This maps characters an equivalent sans-serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSansSerifNoBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold and in the given italics type wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSansSerifNoBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, not in bold and in a italics type. This maps characters an equivalent sans-serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSansSerifBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold with the given italics type wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSansSerifBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol without serifs, in bold and in a italics type. This maps characters an equivalent sans-serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSansSerifNoItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and not in italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSansSerifNoItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and not in italics.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and not in italics. This maps characters an equivalent sans-serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSansSerifItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and in italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSansSerifItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, depending on the given Emphasis in bold or not, and in italics.

Convert the given character to a mathematical symbol without serifs, with a given emphasis and in italics. This maps characters an equivalent sans-serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSansSerifNoBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold, and no italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSansSerifNoBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and not in italics.

Convert the given character to a mathematical symbol without serifs, with no bold and no italics. This maps characters to itself for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSansSerifBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold, and no italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSansSerifBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and not in italics.

Convert the given character to a mathematical symbol without serifs, in bold not in italics. This maps characters an equivalent sans-serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSansSerifNoBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, with no bold, and in italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSansSerifNoBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, not in bold and in italics.

Convert the given character to a mathematical symbol without serifs, with no bold and in italics. This maps characters an equivalent sans-serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSansSerifBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in bold, and in italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSansSerifBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted without serifs, in bold and in italics.

Convert the given character to a mathematical symbol without serifs, with in bold and in italics. This maps characters an equivalent sans-serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.