Data.Char.Math.Script

Description

See Data.Char.Math for further documentation.

Documentation

script

Arguments

:: Emphasis
The given Emphasis style to use.

-> Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is not written in the given Emphasis style. If the character is not supported (see: Data.Char.Math), Nothing is returned.

script'

Arguments

:: Emphasis
The given Emphasis style to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, and depending on the Emphasis in bold or not.

Convert the given character to its script or calligraphic symbol. This symbol is written in the given Emphasis style. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

scriptRegular

Arguments

:: Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is not written in boldface. If the character is not supported (see: Data.Char.Math), Nothing is returned.

scriptRegular'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, not in bold.

Convert the given character to its script or calligraphic symbol. This symbol is not written in boldface. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

scriptBold

Arguments

:: Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is written in boldface. If the character is not supported (see: Data.Char.Math), Nothing is returned.

scriptBold'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, and in bold.

Convert the given character to its script or calligraphic symbol. This symbol is written in boldface. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified.

calligraphy

Arguments

:: Emphasis
The given Emphasis style to use.

-> Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is not written in the given Emphasis style. If the character is not supported (see: Data.Char.Math), Nothing is returned. This is an alias of script.

calligraphy'

Arguments

:: Emphasis
The given Emphasis style to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, and depending on the Emphasis in bold or not.

Convert the given character to its script or calligraphic symbol. This symbol is written in the given Emphasis style. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified. This is an alias of script'.

calligraphyRegular

Arguments

:: Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is not written in boldface. If the character is not supported (see: Data.Char.Math), Nothing is returned. This is an alias of scriptRegular.

calligraphyRegular'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, not in bold.

Convert the given character to its script or calligraphic symbol. This symbol is not written in boldface. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified. This is an alias of scriptRegular'.

calligraphyBold

Arguments

:: Char
The given character to convert.

-> Maybe Char
The calligraphy symbol for the given character wrapped in a Just data constructor, Nothing if there is no equivalent calligraphy character.

Convert the given character to its script or calligraphic symbol wrapped in a Just data constructor. This symbol is written in boldface. If the character is not supported (see: Data.Char.Math), Nothing is returned. This is an alias of scriptBold.

calligraphyBold'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted in calligraphy, and in bold.

Convert the given character to its script or calligraphic symbol. This symbol is written in boldface. If the symbol is not supported (see: Data.Char.Math), the returned character is unspecified. This is an alias of scriptBold'.