Data.Char.Math.Serif

Description

See Data.Char.Math for further documentation.

Serif mathematical alphabet symbols

serif

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol without serifs, in the given emphasis and in the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serif'

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and a given italics style. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifNoBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold and in the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifNoBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, not in bold and in a italics type. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold with the given italics type wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, in bold and in a italics type. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifNoItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in the given emphasis and not in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifNoItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and not in italics.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and not in italics. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in the given emphasis and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and in italics.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and in italics. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifNoBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold, and no italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifNoBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and not in italics.

Convert the given character to a mathematical symbol with serifs, with no bold and no italics. This maps characters to itself for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold, and no italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and not in italics.

Convert the given character to a mathematical symbol with serifs, in bold not in italics. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifNoBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold, and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifNoBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and in italics.

Convert the given character to a mathematical symbol with serifs, with no bold and in italics. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

serifBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold, and in italics wrapped in a Just. If the character is not supported (see: Data.Char.Math), Nothing is returned.

serifBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and in italics.

Convert the given character to a mathematical symbol with serifs, with in bold and in italics. This maps characters an equivalent serif symbol for supported characters (see: Data.Char.Math). For characters outside the range, the behavior is unspecified.

Digit characters

digitSerif

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Maybe Char
The corresponding symbol in serifs for the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character with the given Emphasis in serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Char
The corresponding symbol in serifs for the given emphasis style, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character with a given Emphasis in serif style. The result for characters outside this range is unspecified.

digitSerifRegular

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSerifRegular'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in serifs not in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold serif style. The result for characters outside this range is unspecified.

digitSerifBold

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a bold serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSerifBold'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in serifs in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a bold serif style. The result for characters outside this range is unspecified.

Int to digit characters

intToDigitSerif

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert

-> Maybe Char
The corresponding symbol in serifs in the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character with the given Emphasis in serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert.

-> Char
The corresponding symbol in serifs in the given emphasis style, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character with a given Emphasis in serif style. The result for numbers outside this range is unspecified.

intToDigitSerifRegular

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in a non-bold serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSerifRegular'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding symbol in serifs not in bold, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character in a non-bold serif style. The result for numbers outside this range is unspecified.

intToDigitSerifBold

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in a bold serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSerifBold'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding symbol in serifs in bold, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character in a bold serif style. The result for numbers outside this range is unspecified.