Data.Char.Math.Serif.Digit

Description

See Data.Char.Math for further documentation.

Characters conversion

digitSerif

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Maybe Char
The corresponding symbol in serifs for the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character with the given Emphasis in serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Char
The given character to convert.

-> Char
The corresponding symbol in serifs for the given emphasis style, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character with a given Emphasis in serif style. The result for characters outside this range is unspecified.

digitSerifRegular

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSerifRegular'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in serifs not in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a non-bold serif style. The result for characters outside this range is unspecified.

digitSerifBold

Arguments

:: Char
The given character to convert.

-> Maybe Char
The corresponding symbol in serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given digit character (0–9) to its corresponding character in a bold serif style wrapped in a Just data constructor. For characters outside this range, Nothing is returned.

digitSerifBold'

Arguments

:: Char
The given character to convert.

-> Char
The corresponding symbol in serifs in bold, unspecified outside the the range.

Convert the given digit character (0–9) to its corresponding character in a bold serif style. The result for characters outside this range is unspecified.

Int to digit characters

intToDigitSerif

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert

-> Maybe Char
The corresponding symbol in serifs in the given emphasis style wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character with the given Emphasis in serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSerif'

Arguments

:: Emphasis
The given emphasis style.

-> Int
The given number to convert.

-> Char
The corresponding symbol in serifs in the given emphasis style, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character with a given Emphasis in serif style. The result for numbers outside this range is unspecified.

intToDigitSerifRegular

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in serifs not in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in a non-bold serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSerifRegular'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding symbol in serifs not in bold, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character in a non-bold serif style. The result for numbers outside this range is unspecified.

intToDigitSerifBold

Arguments

:: Int
The given number to convert.

-> Maybe Char
The corresponding symbol in serifs in bold wrapped in a Just, Nothing if the character is outside the range.

Convert the given number (0–9) to its corresponding character in a bold serif style wrapped in a Just data constructor. For numbers outside this range, Nothing is returned.

intToDigitSerifBold'

Arguments

:: Int
The given number to convert.

-> Char
The corresponding symbol in serifs in bold, unspecified outside the the range.

Convert the given number (0–9) to its corresponding character in a bold serif style. The result for numbers outside this range is unspecified.