Data.Char.Math.Serif.Greek

Description

See Data.Char.Math for further documentation.

Documentation

greekSerif

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in the given emphasis and in the given italics type wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSerif'

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and a given italics style. This maps characters an equivalent serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSerifNoBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold and in the given italics type wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSerifNoBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, not in bold and in a italics type. This maps characters an equivalent serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSerifBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold with the given italics type wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSerifBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, in bold and in a italics type. This maps characters an equivalent serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSerifNoItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in the given emphasis and not in italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSerifNoItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and not in italics.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and not in italics. This maps characters an equivalent serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSerifItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in the given emphasis and in italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSerifItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and in italics.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and in italics. This maps characters an equivalent serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSerifNoBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold, and no italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSerifNoBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and not in italics.

Convert the given character to a mathematical symbol with serifs, with no bold and no italics. This maps characters to itself.

greekSerifBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold, and no italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSerifBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and not in italics.

Convert the given character to a mathematical symbol with serifs, in bold not in italics. This maps characters an equivalent serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSerifNoBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold, and in italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSerifNoBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and in italics.

Convert the given character to a mathematical symbol with serifs, with no bold and in italics. This maps characters an equivalent serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.

greekSerifBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold, and in italics wrapped in a Just. If the character is not in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ, Nothing is returned.

greekSerifBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and in italics.

Convert the given character to a mathematical symbol with serifs, with in bold and in italics. This maps characters an equivalent serif symbol for the characters in ΑΒΓΔΕΖΗΘΙΚΛΜΝΞΟΠΡΣΤΥΦΧΨΩ∇ϴαβγδεζηθικλμνξοπρςστυφχψω∂ϵϑϰϕϱϖ. For characters outside the range, the behavior is unspecified.