Data.Char.Math.Serif.Latin

Description

See Data.Char.Math for further documentation.

Documentation

latinSerif

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in the given emphasis and in the given italics type wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSerif'

Arguments

:: ItalicType
The given ItalicType to use.

-> Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and a given italics style. This maps characters an equivalent serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSerifNoBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold and in the given italics type wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSerifNoBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, not in bold and in a italics type. This maps characters an equivalent serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSerifBold

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold with the given italics type wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSerifBold'

Arguments

:: ItalicType
The given ItalicType to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and depending on the given ItalicType in italics or not.

Convert the given character to a mathematical symbol with serifs, in bold and in a italics type. This maps characters an equivalent serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSerifNoItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in the given emphasis and not in italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSerifNoItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and not in italics.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and not in italics. This maps characters an equivalent serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSerifItalic

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in the given emphasis and in italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSerifItalic'

Arguments

:: Emphasis
The given Emphasis to use.

-> Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, depending on the given Emphasis in bold or not, and in italics.

Convert the given character to a mathematical symbol with serifs, with a given emphasis and in italics. This maps characters an equivalent serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSerifNoBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold, and no italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSerifNoBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and not in italics.

Convert the given character to a mathematical symbol with serifs, with no bold and no italics. This maps characters to itself for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSerifBoldNoItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold, and no italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSerifBoldNoItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and not in italics.

Convert the given character to a mathematical symbol with serifs, in bold not in italics. This maps characters an equivalent serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSerifNoBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, with no bold, and in italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSerifNoBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, not in bold and in italics.

Convert the given character to a mathematical symbol with serifs, with no bold and in italics. This maps characters an equivalent serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.

latinSerifBoldItalic

Arguments

:: Char
The given character to convert.

-> Maybe Char
The equivalent character wrapped in a Just if in the valid range, Nothing otherwise.

Convert the given character to a mathematical symbol with serifs, in bold, and in italics wrapped in a Just. If the character is outside the A–Z and a–z range, Nothing is returned.

latinSerifBoldItalic'

Arguments

:: Char
The given character to convert.

-> Char
The equivalent character that is formatted with serifs, in bold and in italics.

Convert the given character to a mathematical symbol with serifs, with in bold and in italics. This maps characters an equivalent serif symbol for the A–Z and a–z range. For characters outside the range, the behavior is unspecified.