Data.Char.Number.Egyptian

Description

One can make use of a block of Unicode characters to typeset ancient Egyptian hieroglyphs. This module aims to make it more convenient to work with ancient Egyptian numerals.

Ancient Egyptian numerals use a sign-value system with radix 10. This means that there are hieroglyphs for 1 to 9, for 10, 20, ..., 90, for 100, 200, ..., 900, etc. With the heh as largest value denoting one million.

In Egyptian numerals, one could write numbers left-to-right and right-to-left (or even vertical). The human or animal figures always point to the beginning of the line. Because in the Unicode standard the heh (denoting one million) looks to the left, the values are written left-to-right.

Ancient plus and minus

plus

Arguments

:: Char
The ancient Egyptian character to denote addition.

The hieroglyph used as a plus sign. This character is used for addition. This is hieroglyph D055.

minus

Arguments

:: Char
The ancient Egyptian character to denote subtraction.

The hieroglyph used as a minus sign. This character is used for subtraction. This is hieroglyph D054.

Ancient Egyptian characters for powers of 10

nfr

Arguments

:: Char
The ancient Egyptian character to denote zero.

The nfr character, an ancient Egyptian character to denote zero, it also means beautiful. This is hieroglyph F035.

singleStroke

Arguments

:: Char
The ancient Egyptian character to denote one.

The single stroke character, an ancient Egyptian character to denote one. This is hieroglyph Z015.

cattleHobble

Arguments

:: Char
The ancient Egyptian character to denote ten.

The cattle hobble character, an ancient Egyptian character to denote ten. This is hieroglyph V020.

coilOfRope

Arguments

:: Char
The ancient Egyptian character to denote hundred.

The coil of rope character, an ancient Egyptian character to denote hundred. This is hieroglyph V001.

waterLily

Arguments

:: Char
The ancient Egyptian character to denote thousand.

The water lily character, an ancient Egyptian character to denote thousand. This is hieroglyph M012.

bentFinger

Arguments

:: Char
The ancient Egyptian character to denote ten thousand.

The bent finger character, an ancient Egyptian character to denote ten thousand. This is hieroglyph B050.

tadpole

Arguments

:: Char
The ancient Egyptian character to denote hundred thousand.

The tadpole character, an ancient Egyptian character to denote hundred thousand. This is hieroglyph I008.

heh

Arguments

:: Char
The ancient Egyptian character to denote one million, or many.

The heh character, an ancient Egyptian character to denote one million, or many. This is hieroglyph C011.

Ancient Egyptian characters for numbers

singleStroke1

Arguments

:: Char
The ancient Egyptian character to denote one.

The single stroke character, an ancient Egyptian character to denote one. This is hieroglyph Z015.

singleStroke2

Arguments

:: Char
The ancient Egyptian character to denote two.

The double stroke character, an ancient Egyptian character to denote two. This is hieroglyph Z015A.

singleStroke3

Arguments

:: Char
The ancient Egyptian character to denote three.

The triple stroke character, an ancient Egyptian character to denote three. This is hieroglyph Z015B.

singleStroke4

Arguments

:: Char
The ancient Egyptian character to denote four.

The quadruple stroke character, an ancient Egyptian character to denote four. This is hieroglyph Z015C.

singleStroke5

Arguments

:: Char
The ancient Egyptian character to denote five.

The quintuple stroke character, an ancient Egyptian character to denote five. This is hieroglyph Z015D.

singleStroke6

Arguments

:: Char
The ancient Egyptian character to denote six.

The sextuple stroke character, an ancient Egyptian character to denote six. This is hieroglyph Z015E.

singleStroke7

Arguments

:: Char
The ancient Egyptian character to denote seven.

The septuple stroke character, an ancient Egyptian character to denote seven. This is hieroglyph Z015F.

singleStroke8

Arguments

:: Char
The ancient Egyptian character to denote eight.

The octuple stroke character, an ancient Egyptian character to denote eight. This is hieroglyph Z015G.

singleStroke9

Arguments

:: Char
The ancient Egyptian character to denote nine.

The nonuple stroke character, an ancient Egyptian character to denote nine. This is hieroglyph Z015H.

singleStroke5'

Arguments

:: Char
The ancient Egyptian character to denote five.

An alternative version of the quintuple stroke character, an ancient Egyptian character to denote five. This is hieroglyph Z015I.

cattleHobble1

Arguments

:: Char
The ancient Egyptian character to denote ten. This is hieroglyph V020.

The cattle hobble character, an ancient Egyptian character to denote ten.

cattleHobble2

Arguments

:: Char
The ancient Egyptian character to denote twenty. This is hieroglyph V020A.

The two cattle hobble character, an ancient Egyptian character to denote twenty.

cattleHobble3

Arguments

:: Char
The ancient Egyptian character to denote thirty.

The three cattle hobble character, an ancient Egyptian character to denote thirty. This is hieroglyph V020B.

cattleHobble4

Arguments

:: Char
The ancient Egyptian character to denote forty.

The four cattle hobble character, an ancient Egyptian character to denote forty. This is hieroglyph V020C.

cattleHobble5

Arguments

:: Char
The ancient Egyptian character to denote fifty. This is hieroglyph V020D.

The five cattle hobble character, an ancient Egyptian character to denote fifty.

cattleHobble6

Arguments

:: Char
The ancient Egyptian character to denote sixty.

The six cattle hobble character, an ancient Egyptian character to denote sixty. This is hieroglyph V020E.

cattleHobble7

Arguments

:: Char
The ancient Egyptian character to denote seventy.

The seven cattle hobble character, an ancient Egyptian character to denote seventy. This is hieroglyph V020F.

cattleHobble8

Arguments

:: Char
The ancient Egyptian character to denote eighty.

The eight cattle hobble character, an ancient Egyptian character to denote eighty. This is hieroglyph V020G.

cattleHobble9

Arguments

:: Char
The ancient Egyptian character to denote ninety.

The nine cattle hobble character, an ancient Egyptian character to denote ninety. This is hieroglyph V020H.

cattleHobble2'

Arguments

:: Char
The ancient Egyptian character to denote twenty.

An alternative version of the two cattle hobble character, an ancient Egyptian character to denote twenty. This is hieroglyph V020I.

cattleHobble3'

Arguments

:: Char
The ancient Egyptian character to denote thirty.

An alternative version of the three cattle hobble character, an ancient Egyptian character to denote thirty. This is hieroglyph V020J.

cattleHobble4'

Arguments

:: Char
The ancient Egyptian character to denote fourty.

An alternative version of the four cattle hobble character, an ancient Egyptian character to denote fourty. This is hieroglyph V020K.

cattleHobble5'

Arguments

:: Char
The ancient Egyptian character to denote fifty.

An alternative version of the five cattle hobble character, an ancient Egyptian character to denote fifty. This is hieroglyph V020L.

coilOfRope1

Arguments

:: Char
The ancient Egyptian character to denote hundred.

The coil of rope character, an ancient Egyptian character to denote hundred. This is hieroglyph V001.

coilOfRope2

Arguments

:: Char
The ancient Egyptian character to denote two hundred.

The double coil of rope character, an ancient Egyptian character to denote two hundred. This is hieroglyph V001A.

coilOfRope3

Arguments

:: Char
The ancient Egyptian character to denote three hundred.

The tripple coil of rope character, an ancient Egyptian character to denote three hundred. This is hieroglyph V001B.

coilOfRope4

Arguments

:: Char
The ancient Egyptian character to denote four hundred.

The quadruple coil of rope character, an ancient Egyptian character to denote four hundred. This is hieroglyph V001C.

coilOfRope5

Arguments

:: Char
The ancient Egyptian character to denote five hundred.

The quintuple coil of rope character, an ancient Egyptian character to denote five hundred. This is hieroglyph V001D.

coilOfRope6

Arguments

:: Char
The ancient Egyptian character to denote six hundred.

The sextuple coil of rope character, an ancient Egyptian character to denote six hundred. This is hieroglyph V001E.

coilOfRope7

Arguments

:: Char
The ancient Egyptian character to denote seven hundred.

The septuple coil of rope character, an ancient Egyptian character to denote seven hundred. This is hieroglyph V001F.

coilOfRope8

Arguments

:: Char
The ancient Egyptian character to denote eight hundred.

The octuple coil of rope character, an ancient Egyptian character to denote eight hundred. This is hieroglyph V001G.

coilOfRope9

Arguments

:: Char
The ancient Egyptian character to denote nine hundred.

The nonuple coil of rope character, an ancient Egyptian character to denote nine hundred. This is hieroglyph V001H.

coilOfRope5'

Arguments

:: Char
The ancient Egyptian character to denote five hundred.

An alternative version of the quintuple coil of rope character, an ancient Egyptian character to denote five hundred. This is hieroglyph V001I.

waterLily1

Arguments

:: Char
The ancient Egyptian character to denote thousand.

The water lily character, an ancient Egyptian character to denote thousand. This is hieroglyph M012.

waterLily2

Arguments

:: Char
The ancient Egyptian character to denote two thousand.

The two water lilies character, an ancient Egyptian character to denote two thousand. This is hieroglyph M012A.

waterLily3

Arguments

:: Char
The ancient Egyptian character to denote three thousand.

The three water lilies character, an ancient Egyptian character to denote three thousand. This is hieroglyph M012B.

waterLily4

Arguments

:: Char
The ancient Egyptian character to denote four thousand.

The four water lilies character, an ancient Egyptian character to denote four thousand. This is hieroglyph M012C.

waterLily5

Arguments

:: Char
The ancient Egyptian character to denote five thousand.

The five water lilies character, an ancient Egyptian character to denote five thousand. This is hieroglyph M012D.

waterLily6

Arguments

:: Char
The ancient Egyptian character to denote six thousand.

The six water lilies character, an ancient Egyptian character to denote six thousand. This is hieroglyph M012E.

waterLily7

Arguments

:: Char
The ancient Egyptian character to denote seven thousand.

The seven water lilies character, an ancient Egyptian character to denote seven thousand. This is hieroglyph M012F.

waterLily8

Arguments

:: Char
The ancient Egyptian character to denote eight thousand.

The eight water lilies character, an ancient Egyptian character to denote eight thousand. This is hieroglyph M012G.

waterLily9

Arguments

:: Char
The ancient Egyptian character to denote nine thousand.

The nine water lilies character, an ancient Egyptian character to denote nine thousand. This is hieroglyph M012H.

bentFinger1

Arguments

:: Char
The ancient Egyptian character to denote ten thousand.

The bent finger character, an ancient Egyptian character to denote ten thousand. This is hieroglyph B050.

bentFinger2

Arguments

:: Char
The ancient Egyptian character to denote twenty thousand.

The two bent fingers character, an ancient Egyptian character to denote twenty thousand. This is hieroglyph B050A.

bentFinger3

Arguments

:: Char
The ancient Egyptian character to denote thirty thousand.

The three bent fingers character, an ancient Egyptian character to denote thirty thousand. This is hieroglyph B050B.

bentFinger4

Arguments

:: Char
The ancient Egyptian character to denote fourty thousand.

The four bent fingers character, an ancient Egyptian character to denote fourty thousand. This is hieroglyph B050C.

bentFinger5

Arguments

:: Char
The ancient Egyptian character to denote fifty thousand.

The five bent fingers character, an ancient Egyptian character to denote fifty thousand. This is hieroglyph B050D.

bentFinger6

Arguments

:: Char
The ancient Egyptian character to denote sixty thousand.

The six bent fingers character, an ancient Egyptian character to denote sixty thousand. This is hieroglyph B050E.

bentFinger7

Arguments

:: Char
The ancient Egyptian character to denote seventy thousand.

The seven bent fingers character, an ancient Egyptian character to denote seventy thousand. This is hieroglyph B050F.

bentFinger8

Arguments

:: Char
The ancient Egyptian character to denote eighty thousand.

The eight bent fingers character, an ancient Egyptian character to denote eighty thousand. This is hieroglyph B050G.

bentFinger9

Arguments

:: Char
The ancient Egyptian character to denote ninety thousand.

The nine bent fingers character, an ancient Egyptian character to denote ninety thousand. This is hieroglyph B050H.

bentFinger5'

Arguments

:: Char
The ancient Egyptian character to denote fifty thousand.

An alternative version of the five bent fingers character, an ancient Egyptian character to denote fifty thousand. This is hieroglyph B050I.

Generating ancient Egyptian numerals

egyptianNumber

Arguments

:: Integral i

=> Ligate
The given ligation style to use.

-> PlusStyle
The given PlusStyle to use.

-> i
The given number to convert to an ancient Egyptian numeral.

-> Text
A sequence of characters that is the ancient Egytian equivalent of the given number.

Construct an ancient Egyptian numeral with the given Ligate style and PlusStyle for the given number.

egyptianNumber'

Arguments

:: Integral i

=> PlusStyle
The given PlusStyle to use.

-> i
The given number to convert to an ancient Egyptian numeral.

-> Text
A sequence of characters that is the ancient Egytian equivalent of the given number.

Construct an acient Egyptian ligated numeral with the given PlusStyle for the given number.

egyptianNumber''

Arguments

:: Integral i

=> i
The given number to convert to an ancient Egyptian numeral.

-> Text
A sequence of characters that is the ancient Egytian equivalent of the given number.

Construct an ancient Egyptian ligated numeral with the default 'PlusStyle.