Data.Char.Number.Mayan

Description

Unicode uses a code block for Mayan numerals. Mayan numerals are written top to bottom, so vertically. This module aims to make it more convenient to write Mayan numerals by offering functions to convert numbers into a Text object for Mayan numbers. Mayan numerals can not present negative numbers.

Define Mayan literals

data MayanLiteral

The Mayan numerals, as defined in the Unicode block.

Constructors

Zero
The unicode character for the Mayan numeral zero: 𝋠.

One
The unicode character for the Mayan numeral one: 𝋡.

Two
The unicode character for the Mayan numeral two: 𝋢.

Three
The unicode character for the Mayan numeral three: 𝋣.

Four
The unicode character for the Mayan numeral four: 𝋤.

Five
The unicode character for the Mayan numeral five: 𝋥.

Six
The unicode character for the Mayan numeral six: 𝋦.

Seven
The unicode character for the Mayan numeral seven: 𝋧.

Eight
The unicode character for the Mayan numeral eight: 𝋨.

Nine
The unicode character for the Mayan numeral nine: 𝋩.

Ten
The unicode character for the Mayan numeral ten: 𝋪.

Eleven
The unicode character for the Mayan numeral eleven: 𝋫.

Twelve
The unicode character for the Mayan numeral twelve: 𝋬.

Thirteen
The unicode character for the Mayan numeral thirteen: 𝋭.

Fourteen
The unicode character for the Mayan numeral fourteen: 𝋮.

Fifteen
The unicode character for the Mayan numeral fifteen: 𝋯.

Sixteen
The unicode character for the Mayan numeral sixteen: 𝋰.

Seventeen
The unicode character for the Mayan numeral seventeen: 𝋱.

Eighteen
The unicode character for the Mayan numeral eighteen: 𝋲.

Nineteen
The unicode character for the Mayan numeral nineteen: 𝋳.

Converting integers to Mayan numbers.

toMayanVertical

Arguments

:: Integral i

=> i
The given number to convert to a vertical String object.

-> Maybe Text
A Text that contains the Mayan number wrapped in a Just if we can represent the number; Nothing otherwise.

Convert the given Integral number to a Text object that writes the Mayan number to to bottom. This function will return a Nothing in case the number is negative (since it can not be presented in Mayan).

toMayanVertical'

Arguments

:: Integral i

=> i
The given number to convert to a vertical String object.

-> Text
A Text that contains the Mayan number.

Convert the given Integral number to a Text object that writes the Mayan number to to bottom.

toMayanVertical''

Arguments

:: Integral i

=> i
The given number to convert to a vertical String object.

-> String
A String that contains the Mayan number.

Convert the given Integral number to a String object that writes the Mayan number to to bottom.

toMayan

Arguments

:: Integral i

=> i
The given number to convert to a horizontal String object.

-> Maybe Text
A Text that contains the Mayan number wrapped in a Just if we can represent the number; Nothing otherwise.

Convert the given Integral number to a Text object that writes the Mayan number from left-to-right. The object is wrapped in a Just data constructor. If the number is negative, and thus can not be represented, Nothing is returned.

toMayan'

Arguments

:: Integral i

=> i
The given number to convert to a horizontal String object.

-> Text
A Text that contains the Mayan number.

Convert the given Integral number to a Text object that writes the Mayan number from left-to-right.

toMayan''

Arguments

:: Integral i

=> i
The given number to convert to a horizontal String object.

-> String
A String that contains the Mayan number.

Convert the given Integral number to a String object that writes the Mayan number from left-to-right.