Data.Char.Number.Segmented

Description

The 1FB00 Unicode block contains symbols for legacy computing and among them segmented digits: 🯰🯲🯳🯴🯵🯶🯷🯸🯹

Convert numbers to a seven-segmented display

segmentedDigit

Arguments

:: Integral i

=> i
The given number to convert.

-> Maybe Char
A Unicode character that represents the given digit on a seven-segment display wrapped in a Just data constructor. If the number is outside the 0-9 range, then Nothing is returned.

Convert the given Integral number to its Unicode character wrapped in a Just data constructor. In case the value is less than 0 or greater than 9, Nothing is returned.

segmentedDigit'

Arguments

:: Integral i

=> i
The given number to convert.

-> Char
A Unicode character that represents this digit.

Convert the given Integral number to its Unicode character. In case the value is less than 0, or greater than 9, the behavior is unspecified.