Data.Char.Number.Tally

Description

This module aims to convert numbers to (Western) tally marks and vice versa.

Data types to represent tally marks

data TallyLiteral

A tally literal that is either a one (𝍷), or five grouped together (𝍸).

Constructors

I
The unicode character for the tally numeral one: 𝍷.

V
The unicode character for the tally numeral five: 𝍸.

Convert a number to TallyLiterals

toLiterals

Arguments

:: Integral i

=> i
The given number to convert.

-> Maybe [TallyLiteral]
A list of TallyLiterals if the given number can be specified with tally marks, Nothing otherwise.

Convert a given positive natural number to a sequence of TallyLiterals.

toLiterals'

Arguments

:: Integral i

=> i
The given number to convert.

-> [TallyLiteral]
A list of TallyLiterals that denotes the given number.

Convert a given number to a sequence of TallyLiterals, for negative numbers or zero, the behavior is unspecified.

tallyNumber

Arguments

:: Integral i

=> i
The given number to convert.

-> Maybe Text
A Text with the tally marks wrapped in a Just if the number can be represented with tally marks; Nothing otherwise.

Convert a given positive natural number to a Text object with the tally marks for that number.

tallyNumber'

Arguments

:: Integral i

=> i
The given number to convert.

-> Text
The corresponding Text that contains the number as tally marks.

Convert a given number to a Text object with the tally marks for that number, for negative numbers or zero, the behavior is unspecified.