Data.Char.Private

Description

Unicode has a block named Private Use Area that can be used for all sorts of characters, and each font can decide to allocate certain characters for this block. In the submodules, popular usages are supoorted.

Character bounds

privateStart

Arguments

:: Char
The first Char in the private use area.

The first character in the private use area.

privateStop

Arguments

:: Char
The last Char in the private use area.

The last character in the private use area.

isPrivateChar

Arguments

:: Char
The given Char to check.

-> Bool
True if the Char is in the private use area; False otherwise.

Checks if the given Char is in the private use area.

Generating private characters

privateCharGen

Arguments

:: Gen Char
A Gen that generates Char objects in the private use area.

A generator of Characters in the private use area.