Data.Char.Private.Klingon

Description

Some fonts implement the Klingon script characters with the private usage area. Wikipedia describes the different fonts and Evertype provides a font to typeset Klingon script.

Representing Klingon characters

data Klingon

A datatype to represent the Klingon characters, and their mapping to a Klingon font.

Constructors

A
The a character in Klingon script.

B
The b character in Klingon script.

Ch
The ch character in Klingon script.

D
The D character in Klingon script.

E
The e character in Klingon script.

Gh
The gh character in Klingon script.

H
The H character in Klingon script.

I
The I character in Klingon script.

J
The j character in Klingon script.

L
The l character in Klingon script.

M
The m character in Klingon script.

N
The n character in Klingon script.

Ng
The ng character in Klingon script.

O
The o character in Klingon script.

P
The p character in Klingon script.

Q
The q character in Klingon script.

QUpper
The Q character in Klingon script.

R
The r character in Klingon script.

S
The S character in Klingon script.

T
The t character in Klingon script.

Tlh
The tlh character in Klingon script.

U
The u character in Klingon script.

V
The v character in Klingon script.

W
The w character in Klingon script.

Y
The y character in Klingon script.

GlottalStop
The ʼ character in Klingon script, denoting the glottal stop.

Zero
The 0 character in Klingon script.

One
The 1 character in Klingon script.

Two
The 2 character in Klingon script.

Three
The 3 character in Klingon script.

Four
The 4 character in Klingon script.

Five
The 5 character in Klingon script.

Six
The 6 character in Klingon script.

Seven
The 7 character in Klingon script.

Eight
The 8 character in Klingon script.

Nine
The 9 character in Klingon script.

Comma
The comma character in Klingon script, denoted by an up-turned triangle.

FullStop
The full stop character in Klingon script, denoted by a down-turned triangle.

Mummification
The mummification character in Klingon script, also known as the klingon character for the empire or heart of virtue.