Data.Char.Block

Description

Unicode has 2-by-2 blocks, this module aims to make it more convenient to render such blocks.

Datastructures to store the state of the frame.

data Row a

A data type that determines the state of the row in a block. it determines the left and the right part of the row of the block.

Constructors

Row

Fields

left :: a

The left part of a row of the block.

right :: a

The right part of the row of the block.

data Block a

A data type that determines the state of the four subparts of the block.

Constructors

Block

Fields

upper :: Row a

The upper part of the block.

lower :: Row a

The lower part of the block.

A unicode character that is (partially) filled block.

filled :: Block Bool -> Char

Convert the given Block value to a block character in unicode. True means that part is filled, and False means the part is not filled.