Data.Char.Core

Description

This module defines data structures that are used in other modules, for example to rotate the characters.

Possible rotations

data Orientation

The possible orientations of a unicode character, these can be horizontal, or vertical.

Constructors

Horizontal
Horizontal orientation.

Vertical
Vertical orientation.

data Rotate90

Possible rotations of a unicode character if that character can be rotated over 0, 90, 180, and 270 degrees.

Constructors

R0
No rotation.

R90
Rotation over 90 degrees.

R180
Rotation over 180 degrees.

R270
Rotation over 270 degrees.

Rotated objects

data Oriented a

A data type that specifies that an item has been given an orientation.

Constructors

Oriented

Fields

oobject :: a

The object that is oriented.

orientation :: Orientation

The oriented of the oriented object.

Letter case

data LetterCase

Specify whether we write a value in UpperCase or LowerCase. The Default is UpperCase, since for example often Roman numerals are written in upper case.

Constructors

UpperCase
The upper case formatting.

LowerCase
The lower case formatting.

splitLetterCase

Arguments

:: a
The value to return in case of UpperCase.

-> a
The value to return in case of LowerCase.

-> LetterCase
The given letter case.

-> a
One of the two given values, depending on the LetterCase value.

Pick one of the two values based on the LetterCase value.

Ligating

data Ligate

Specify if one should ligate, or not. When litigation is done characters that are normally written in two (or more) characters are combined in one character. For example Ⅲ instead of ⅠⅠⅠ.

Constructors

Ligate
A ligate operation is performed on the characters, the def for 't:Ligate'.

NoLigate
No ligate operation is performed on the charaters.

splitLigate

Arguments

:: a
The value to return in case of 'v:Ligate'.

-> a
The value to return in case of NoLigate.

-> Ligate
The ligation style.

-> a
One of the two given values, based on the 't:Ligate' value.

Pick one of the two values based on the value for 't:Ligate'.

ligate :: (a -> a) -> Ligate -> a -> a

Specify if the given ligate function should be performed on the input, if 'v:Ligate' is passed, and the identity function otherwise.

ligateF :: Functor f => (a -> a) -> Ligate -> f a -> f a

Specify if the given ligate function is performed over the functor object if 'v:Ligate' is passed, and the identity function otherwise.

Types of fonts

data Emphasis

A data type that lists the possible emphasis of a font. This can be Bold or NoBold the Default is NoBold.

Constructors

NoBold
The characters are not stressed with boldface.

Bold
The characters are stressed in boldface.

splitEmphasis

Arguments

:: a
The value to return in case of NoBold.

-> a
The value to return in case of Bold.

-> Emphasis
The emphasis type.

-> a
One of the two given values, based on the 't:Emphasis' value.

Pick one of the two values based on the 't:Emphasis' value.

data ItalicType

A data type that can be used to specify if an italic character is used. The Default is NoItalic.

Constructors

NoItalic
No italic characters are used.

Italic
Italic characters are used.

splitItalicType

Arguments

:: a
The value to return in case of NoItalic.

-> a
The value to return in case of Italic.

-> ItalicType
The italic type.

-> a
One of the two given values, based on the 't:ItalicType' value.

Pick one of the two values based on the 't:ItalicType' value.

data FontStyle

A data type that specifies if the font is with serifs or not. The 'Defaul;t' is Serif.

Constructors

SansSerif
The character is a character rendered without serifs.

Serif
The character is a character rendered with serifs.

splitFontStyle

Arguments

:: a
The value to return in case of SansSerif.

-> a
The value to return in case of Serif.

-> FontStyle
The font style.

-> a
One of the two given values, based on the 't:FontStyle' value.

Pick one of the two values based on the 't:FontStyle' value.

Character range checks

isAsciiAlphaNum :: Char -> Bool

Checks if a character is an alphabetic or numerical character in ASCII. The characters 0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz satisfy this predicate.

isAsciiAlpha :: Char -> Bool

Checks if a charcter is an alphabetic character in ASCII. The characters ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz satisfy this predicate.

isACharacter

Arguments

:: Char
The given Character to check.

-> Bool
True if the given Character is a character (according to the Unicode specifications); False otherwise.

Check if the given character is a character according to the Unicode specifications. Codepoints that are not a character are denoted in the Unicode documentation with <not a character>.

isNotACharacter

Arguments

:: Char
The given Character to check.

-> Bool
True if the given Character is not a character (according to the Unicode specifications); False otherwise.

Check if the given character is not a character according to the Unicode specifications. The Unicode documentation denotes these with <not a character>.

isReserved

Arguments

:: Char
The given Character to check.

-> Bool
True if the given Character is reserved; False otherwise.

Check if the given character is a reserved character. This is denoted in the Unicode documentation with <reserved>.

Ways to display numbers

data PlusStyle

Specify whether we write a positive number with or without a plus sign. the Default is WithoutPlus.

Constructors

WithoutPlus
Write positive numbers without using a plus sign.

WithPlus
Write positive numbers with a plus sign.

splitPlusStyle

Arguments

:: a
The value to return in case of WithoutPlus.

-> a
The value to return in case of WithPlus.

-> PlusStyle
The plus style.

-> a
One of the two given values, based on the 't:PlusStyle' value.

Pick one of the two values based on the 't:PlusStyle' value.

Functions to implement a number system

withSign

Arguments

:: Integral i

=> (i -> Text)
The function that maps the absolute value of the number to a Text object that is appended to the sign.

-> Char
The plus sign to use.

-> Char
The minus sign to use.

-> PlusStyle
The given PlusStyle to use.

-> i
The given Integral number to render.

-> Text
A Text object that represents the given number, with the given sign numbers in the given PlusStyle.

Calculate for a given plus and minus sign a Text object for the given number in the given PlusStyle.

signValueSystem

Arguments

:: Integral i

=> i
The given radix to use.

-> (Int -> Int -> Text)
A function that maps the value and the weight to a Text object.

-> Text
The given Text used to represent zero.

-> Char
The given Char used to denote plus.

-> Char
The given Char used to denote minus.

-> PlusStyle
The given PlusStyle to use.

-> i
The given number to convert.

-> Text
A Text object that denotes the given number with the given sign-value system.

A function to make it more convenient to implement a sign-value system. This is done for a given radix a function that maps the given value and the given weight to a Text object, a Text object for zero (since in some systems that is different), and characters for plus and minus. The function then will for a given PlusStyle convert the number to a sequence of characters with respect to how the sign-value system is implemented.

positionalNumberSystem

Arguments

:: Integral i

=> i
The given radix to use.

-> (Int -> Char)
A function that maps the value of a digit to the corresponding Char.

-> Char
The given character used to denote plus.

-> Char
The given character used to denote minus.

-> PlusStyle
The given PlusStyle to use.

-> i
The given number to convert.

-> Text
A Text object that denotes the given number with the given positional number system.

A function to make it more convenient to implement a /positional number system. This is done for a given radix/ a given conversion funtion that maps a value to a Char, and a Char for plus and minus. The function then construct a Text object for a given PlusStyle and a given number.

positionalNumberSystem10

Arguments

:: Integral i

=> (Int -> Char)
A function that maps the value of a digit to the corresponding Char.

-> Char
The given character used to denote plus.

-> Char
The given character used to denote minus.

-> PlusStyle
The given PlusStyle to use.

-> i
The given number to convert.

-> Text
A Text object that denotes the given number with the given positional number system.

A function to make it more convenient to implement a /positional number system with radix/ 10.

Re-export of some functions of the Char module

chr :: Int -> Char

The toEnum method restricted to the type Char.

isAlpha :: Char -> Bool

Selects alphabetic Unicode characters (lower-case, upper-case and title-case letters, plus letters of caseless scripts and modifiers letters). This function is equivalent to isLetter.

isAlphaNum :: Char -> Bool

Selects alphabetic or numeric Unicode characters.

Note that numeric digits outside the ASCII range, as well as numeric characters which aren't digits, are selected by this function but not by isDigit. Such characters may be part of identifiers but are not used by the printer and reader to represent numbers.

isAscii :: Char -> Bool

Selects the first 128 characters of the Unicode character set, corresponding to the ASCII character set.

ord :: Char -> Int

The fromEnum method restricted to the type Char.