Data.Char.Core

Description

This module defines data structures that are used in other modules, for example to rotate the characters.

Possible rotations

data Orientation

The possible orientations of a unicode character, these can be horizontal, or vertical.

Constructors

Horizontal
Horizontal orientation.

Vertical
Vertical orientation.

data Rotate90

Possible rotations of a unicode character if that character can be rotated over 0, 90, 180, and 270 degrees.

Constructors

R0
No rotation.

R90
Rotation over 90 degrees.

R180
Rotation over 180 degrees.

R270
Rotation over 270 degrees.

Rotated objects

data Oriented a

A data type that specifies that an item has been given an orientation.

Constructors

Oriented

Fields

oobject :: a

The object that is oriented.

orientation :: Orientation

The oriented of the oriented object.

Ligating

data Ligate

Specify if one should ligate, or not. When litigation is done characters that are normally written in two (or more) characters are combined in one character. For example Ⅲ instead of ⅠⅠⅠ.

Constructors

Ligate
A ligate operation is performed on the characters.

NoLigate
No ligate operation is performed on the charaters.

ligate :: (a -> a) -> Ligate -> a -> a

Specify if the given ligate function should be performed on the input, if 'v:Ligate' is passed, and the identity function otherwise.

ligateF :: Functor f => (a -> a) -> Ligate -> f a -> f a

Specify if the given ligate function is performed over the functor object if 'v:Ligate' is passed, and the identity function otherwise.

Types of fonts

data Emphasis

A data type that lists the possible emphasis of a font. This can be Bold or NoBold the Default is NoBold.

Constructors

NoBold
The characters are not stressed with boldface.

Bold
The characters are stressed in boldface.

data ItalicType

A data type that can be used to specify if an italic character is used. The Default is NoItalic.

Constructors

NoItalic
No italic characters are used.

Italic
Italic characters are used.

data FontStyle

A data type that specifies if the font is with serifs or not. The 'Defaul;t' is Serif.

Constructors

SansSerif
The character is a character rendered without serifs.

Serif
The character is a character rendered with serifs.

Character range checks

isAsciiAlphaNum :: Char -> Bool

Checks if a character is an alphabetic or numerical character in ASCII. The characters 0123456789ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz satisfy this predicate.

isAsciiAlpha :: Char -> Bool

Checks if a charcter is an alphabetic character in ASCII. The characters ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz satisfy this predicate.