Data.Char.Core

Description

This module defines data structures that are used in other modules, for example to rotate the characters.

Possible rotations

data Orientation

The possible orientations of a unicode character, these can be horizontal, or vertical.

Constructors

Horizontal
Horizontal orientation.

Vertical
Vertical orientation.

data Rotate90

Possible rotations of a unicode character if that character can be rotated over 0, 90, 180, and 270 degrees.

Constructors

R0
No rotation.

R90
Rotation over 90 degrees.

R180
Rotation over 180 degrees.

R270
Rotation over 270 degrees.

Rotated objects

data Oriented a

A data type that specifies that an item has been given an orientation.

Constructors

Oriented

Fields

oobject :: a

The object that is oriented.

orientation :: Orientation

The oriented of the oriented object.

Ligating

data Ligate

Specify if one should ligate, or not. When litigation is done characters that are normally written in two (or more) characters are combined in one character. For example Ⅲ instead of ⅠⅠⅠ.

Constructors

Ligate
A ligate operation is performed on the characters.

NoLigate
No ligate operation is performed on the charaters.

ligate :: (a -> a) -> Ligate -> a -> a

Specify if the given ligate function should be performed on the input, if 'v:Ligate' is passed, and the identity function otherwise.

ligateF :: Functor f => (a -> a) -> Ligate -> f a -> f a

Specify if the given ligate function is performed over the functor object if 'v:Ligate' is passed, and the identity function otherwise.