Data.Char.Frame

Description

A frame is represented as a pair of horizontal and vertical lines. These can be any items, but currently only booleans and weight objects are covered to convert the item to a corresponding character.

Line weight

data Weight

The weights of the frame lines, these can be Empty, Light or Heavy.

Constructors

Empty
The frame does not contain such line.

Light
The frame contains such line.

Heavy
The frame contains such line, in boldface.

Datastructures to store the four directions

data Horizontal a

A data type that determines the state of the horizontal lines of the frame (left and right).

Constructors

Horizontal

Fields

left :: a

The state of the left line of the frame.

right :: a

The state of the right line of the frame.

data Vertical a

A data type that determines the state of the vertical lines of the frame (up and down).

Constructors

Vertical

Fields

up :: a

The state of the line in the up direction of the frame.

down :: a

The state of the line in the down direction of the frame.

data Parts a

A data type that specifies the four lines that should (not) be drawn for the frame.

Constructors

Parts (Vertical a) (Horizontal a)

Type aliasses and pattern synonyms for convenient Parts

type Simple = Parts Bool

A type synonym that makes it more convenient to work with a Parts object that wraps Bools. Usually True means it should draw a line, and False that there is no line in that direction.

type Weighted = Parts Weight

A type synonym that makes it more convenient to work with a Parts object that wraps Weight objects. These specify the weight .

pattern Frame

Arguments

:: a
The state of the line in the up direction.

-> a
The state of the line in the down direction.

-> a
The state of the line in the left direction.

-> a
The state of the line in the right direction.

-> Parts a
The Parts pattern with the state of the given lines.

A pattern that makes pattern matching and expressions with Parts more convenient.

Functions to render specific frame values

simple

Arguments

:: Simple
The given Simple frame to convert.

-> Char
The corresponding characer for this Simple frame.

Convert a Simple frame to a corresponding Char. Here True is mapped to a Light line.

simple'

Arguments

:: Simple
The given Simple frame to convert.

-> Char
The corresponding characer for this Simple frame.

Convert a Simple frame to a corresponding Char. Here True is mapped to a Heavy line.

simpleWithArc

Arguments

:: Simple
The given Simple frame to convert.

-> Char
The corresponding characer for this Simple frame.

Generate a Char where turns are done with an arc instead of a corner. This can only be done for Light lines.

weighted

Arguments

:: Weighted
The Weighted object that specifies how the lines on the four directions should look like.

-> Char
The character that represents these lines.

Converts a given Weighted to the char that can be used to render frames.

Convert a Simple to a Weighted

simpleToWeighted

Arguments

:: Weight
The Weight that is used for True values.

-> Simple
The Simple frame to convert.

-> Weighted
The resulting Weighted frame.

Convert a Simple frame to a Weighted frame by converting True to the given Weight value.

simpleToLight

Arguments

:: Simple
The Simple frame to convert.

-> Weighted
The resulting Weighted frame.

Convert a Simple frame to a Weighted frame by converting True to Light.

simpleToHeavy

Arguments

:: Simple
The 'Simple frame to convert.

-> Weighted
The resulting Weighted frame.

Convert a Simple frame to a Weighted frame by converting True to Heavy.